Name of last commentator

Retrieve the name of the last commentator

author_last_comment(x, ...)

# S3 method for class 'IssueTB'
author_last_comment(x, verbose = TRUE, ...)

# S3 method for class 'IssuesTB'
author_last_comment(x, verbose = TRUE, ...)

Arguments

x

An object of class IssueTB or IssuesTB.

...

Currently not used.

verbose

A boolean indicating whether to print additional information. Default is TRUE.

Value

A string with the name of the last person which leaves a comment. If there is no comments, it returns an empty string.

Examples

issues_selector <- init_selector(
    source = "Local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = NULL
)
my_issues <- get_issues(selector = issues_selector)
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_issues.yaml.
author_last_comment(my_issues)
#> Try to retrieve comments from the list of issues.
#>   [1] "charphi"             ""                    "TanguyBarthelemy"   
#>   [4] ""                    "palatej"             "charphi"            
#>   [7] ""                    ""                    ""                   
#>  [10] ""                    "TanguyBarthelemy"    ""                   
#>  [13] ""                    ""                    "TanguyBarthelemy"   
#>  [16] ""                    ""                    ""                   
#>  [19] ""                    "annasmyk"            "charphi"            
#>  [22] "charphi"             ""                    "charphi"            
#>  [25] "palatej"             "TanguyBarthelemy"    "Blassik"            
#>  [28] ""                    ""                    "TanguyBarthelemy"   
#>  [31] "TanguyBarthelemy"    "TanguyBarthelemy"    "TanguyBarthelemy"   
#>  [34] "TanguyBarthelemy"    "palatej"             "OlesjaL"            
#>  [37] ""                    ""                    ""                   
#>  [40] "TanguyBarthelemy"    ""                    "Immurb"             
#>  [43] ""                    "charphi"             ""                   
#>  [46] "Immurb"              "TanguyBarthelemy"    "TanguyBarthelemy"   
#>  [49] "TanguyBarthelemy"    ""                    "charphi"            
#>  [52] ""                    ""                    ""                   
#>  [55] "charphi"             ""                    ""                   
#>  [58] ""                    "charphi"             "charphi"            
#>  [61] "charphi"             "charphi"             ""                   
#>  [64] ""                    ""                    ""                   
#>  [67] "a-s-russo"           ""                    "palatej"            
#>  [70] "palatej"             "Immurb"              "charphi"            
#>  [73] ""                    "AQLT"                ""                   
#>  [76] "TanguyBarthelemy"    "TanguyBarthelemy"    ""                   
#>  [79] ""                    "AQLT"                "TanguyBarthelemy"   
#>  [82] ""                    "charphi"             "TanguyBarthelemy"   
#>  [85] "palatej"             "TanguyBarthelemy"    ""                   
#>  [88] "AlessandroPiovani"   ""                    "TanguyBarthelemy"   
#>  [91] ""                    ""                    "palatej"            
#>  [94] "palatej"             "charphi"             "palatej"            
#>  [97] ""                    ""                    ""                   
#> [100] ""                    "MichelleHedger"      "charphi"            
#> [103] "charphi"             ""                    "TanguyBarthelemy"   
#> [106] "annasmyk"            ""                    ""                   
#> [109] ""                    ""                    ""                   
#> [112] "charphi"             "TanguyBarthelemy"    "palatej"            
#> [115] "TanguyBarthelemy"    "palatej"             "charphi"            
#> [118] ""                    "a-s-russo"           "charphi"            
#> [121] "charphi"             "palatej"             "palatej"            
#> [124] "MichelleHedger"      "annasmyk"            ""                   
#> [127] ""                    "pmbyron"             ""                   
#> [130] "palatej"             "annasmyk"            "annasmyk"           
#> [133] ""                    ""                    ""                   
#> [136] "annasmyk"            ""                    ""                   
#> [139] "carlos-saez"         "carlos-saez"         "palatej"            
#> [142] "a-s-russo"           ""                    "palatej"            
#> [145] "palatej"             "palatej"             "TanguyBarthelemy"   
#> [148] "Ciamman"             "palatej"             ""                   
#> [151] ""                    "a-s-russo"           "charphi"            
#> [154] ""                    "charphi"             "charphi"            
#> [157] "TanguyBarthelemy"    "charphi"             "a-s-russo"          
#> [160] "a-s-russo"           "palatej"             "CyLuGh"             
#> [163] "charphi"             ""                    "pmbyron"            
#> [166] "palatej"             "palatej"             ""                   
#> [169] "Immurb"              "Immurb"              ""                   
#> [172] "TanguyBarthelemy"    "TanguyBarthelemy"    ""                   
#> [175] "MichelleHedger"      "Immurb"              ""                   
#> [178] "a-s-russo"           "walkman-w902"        "a-s-russo"          
#> [181] ""                    "Immurb"              "palatej"            
#> [184] "palatej"             "AQLT"                "charphi"            
#> [187] "Immurb"              "Immurb"              ""                   
#> [190] "MFatihTuzen"         ""                    "charphi"            
#> [193] ""                    "a-s-russo"           "a-s-russo"          
#> [196] "palatej"             "AQLT"                ""                   
#> [199] ""                    ""                    "palatej"            
#> [202] ""                    "a-s-russo"           "a-s-russo"          
#> [205] "Immurb"              "charphi"             "a-s-russo"          
#> [208] "Immurb"              ""                    "Immurb"             
#> [211] "palatej"             ""                    "a-s-russo"          
#> [214] ""                    "charphi"             "a-s-russo"          
#> [217] "charphi"             "TanguyBarthelemy"    "charphi"            
#> [220] ""                    "charphi"             ""                   
#> [223] "palatej"             "palatej"             "charphi"            
#> [226] ""                    "charphi"             "charphi"            
#> [229] "palatej"             ""                    "palatej"            
#> [232] "TanguyBarthelemy"    ""                    ""                   
#> [235] "charphi"             "palatej"             "TanguyBarthelemy"   
#> [238] ""                    "mr-abdullah-shahzad" "a-s-russo"          
#> [241] "charphi"             "charphi"             "charphi"            
#> [244] "palatej"             "m20l22"              "palatej"            
#> [247] "charphi"             "charphi"             "TanguyBarthelemy"   
#> [250] "charphi"             ""                    ""                   
#> [253] "charphi"             "a-s-russo"           "charphi"            
#> [256] "palatej"             "charphi"             "charphi"            
#> [259] "charphi"             "charphi"             "charphi"            
#> [262] "palatej"             "ChristianeHofer"     "palatej"            
#> [265] "charphi"             "charphi"             ""                   
#> [268] "charphi"             "TanguyBarthelemy"    "charphi"            
#> [271] "charphi"             "AQLT"                "palatej"            
#> [274] "a-s-russo"           "AQLT"                ""                   
#> [277] "charphi"             "palatej"             "charphi"            
#> [280] "charphi"             "TanguyBarthelemy"    "TanguyBarthelemy"   
#> [283] "TanguyBarthelemy"    "palatej"             "ChristianeHofer"    
#> [286] "ChristianeHofer"    
author_last_comment(my_issues[1L, ])
#> [1] "charphi"