Export an R Object to YAML

This function exports an R object (such as a list, vector, data.frame, etc.) to a YAML file.

.write(
  x,
  dataset_dir = tempdir(),
  dataset_name = "object.yaml",
  overwrite = TRUE,
  verbose = TRUE,
  ...
)

Arguments

x

An R object to export.

dataset_dir

The destination directory where the YAML file will be saved. By default, the system's temporary directory is used (tempdir()).

dataset_name

The name of the output file (without extension). By default, the name is "object.yaml".

overwrite

Logical indicating whether to overwrite the file if it already exists. Defaults to TRUE.

verbose

A logical value indicating whether to print additional information. Default is TRUE.

...

Currently not used.

Value

The function returns invisibly the full path of the written YAML file. If the file already exists and overwrite = FALSE, it returns FALSE without writing.

Details

The function automatically handles directory creation when the path doesn't exist.

Examples

my_list <- list(name = "John", age = 30, city = "Paris")
IssueTrackeR:::.write(my_list, dataset_name = "example_list")
#> The datasets will be exported to /tmp/RtmpS8xnmK/example_list.yaml.

my_df <- data.frame(id = 1:3, value = c("A", "B", "C"))
my_data_dir <- tempfile("data")
IssueTrackeR:::.write(
    x = my_df,
    dataset_dir = my_data_dir,
    dataset_name = "my_dataframe"
)
#> The datasets will be exported to /tmp/RtmpS8xnmK/data1a70865f25/my_dataframe.yaml.

IssueTrackeR:::.write(
    x = my_list,
    dataset_name = "example_list",
    overwrite = FALSE
)
#> The file already exists and won't be overwritten. To overwrite this file, please set `overwrite = TRUE`.