Retrieve information from the issues of GitHub

use gh to ask the API of GitHub and et a list of issues with their labels and milestones.

get_issues(selector = getOption("IssueTrackeR.selector"), verbose = TRUE, ...)

get_labels(selector = getOption("IssueTrackeR.selector"), verbose = TRUE, ...)

get_milestones(
  selector = getOption("IssueTrackeR.selector"),
  verbose = TRUE,
  ...
)

Arguments

selector

A SelectorTB object to check. See the corresponding vignettes: https://tanguybarthelemy.github.io/IssueTrackeR/articles/selectors.html to have more information about selectors. By default getOption("IssueTrackeR.selector").

verbose

A boolean indicating whether to print additional information. Default is TRUE.

...

Additional arguments to connect to the GitLab as gitlab_url and private_token. (See the documentation of get to have more information on theses parameters)

Value

The function get_issues returns an object of class IssuesTB. It is a list composed by object of class IssueTB. An object of class IssueTB represents an issue with simpler structure (with number, title, body and labels).

The function get_labels returns a list representing labels with simpler structure (with name, description, colour).

The function get_milestones returns a list representing milestones with simpler structure (with title, description and due_on).

Details

The functions of get type are useful to retrieve object related to issues from GitHub. So it's possible to retrieve issues, labels and milestones.

The defaults value for the argument dataset_name depends on the function:

defaults is "list_issues.yaml" for get_issues()

defaults is "list_milestones.yaml" for get_milestones()

defaults is "list_labels.yaml" for get_labels()

Examples

# \donttest{
# From online

online_selector <- init_selector(
    source = "GitHub",
    owner = "TanguyBarthelemy",
    repo = "IssueTrackeR"
)
issues <- get_issues(selector = online_selector)
#> Repo: IssueTrackeR  owner: TanguyBarthelemy 
#> Error in assert_issue(title = title, body = body, number = number, state = state,     created_at = created_at, closed_at = closed_at, closed_by = closed_by,     labels = labels, milestone = milestone, repo = repo, url = url,     owner = owner, html_url = html_url, comments = comments,     creator = creator, assignee = assignee, state_reason = state_reason): Assertion on 'labels' failed: Must be of type 'data.frame' (or 'NULL'), not 'list'.
print(issues)
#> Error: object 'issues' not found

labels <- get_labels(selector = online_selector)
#> Repo: IssueTrackeR  owner: TanguyBarthelemy 
#> Reading labels... Done!
#> 41 labels found.
print(labels)
#> There is 1 repo.
#> - TanguyBarthelemy/IssueTrackeR:, automated issue, communication, complexity: complex, complexity: moderate, complexity: trivial/easy, complexity: very-complex, documentation, duplicate, effort: long, effort: medium, effort: short, effort: unknown, effort: very-long, good first issue, good practice, help wanted, invalid, missing examples, priority: critical, priority: high, priority: low, priority: medium, question, report, status: can't-reproduce, status: completed, status: duplicate, status: in-progress, status: needs-information, status: on-hold, status: review-needed, status: triage, status: won't-do/fix, tests, type: bug, type: enhancement, type: feature, type: maintenance, type: question, typo, wontfix

milestones <- get_milestones(selector = online_selector)
#> Repo: IssueTrackeR  owner: TanguyBarthelemy 
#> Reading milestones... 
#> 	-  v2.1.0 ... Done!
#> 	-  v2.0.0 ... Done!
#> Done! 2 milestones found.
print(milestones)
#>   source  title description     due_on closed_at          creator state
#> 1 GitHub v2.1.0                   <NA>      <NA> TanguyBarthelemy  open
#> 2 GitHub v2.0.0             2026-10-31      <NA> TanguyBarthelemy  open
#>   nb_issues_open nb_issues_closed         repo            owner
#> 1             10                0 IssueTrackeR TanguyBarthelemy
#> 2              3               22 IssueTrackeR TanguyBarthelemy
#>                                                            url
#> 1 https://github.com/TanguyBarthelemy/IssueTrackeR/milestone/7
#> 2 https://github.com/TanguyBarthelemy/IssueTrackeR/milestone/2
# }

# From local

local_selector <- init_selector(
    source = "Local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = NULL
)

issues <- get_issues(selector = local_selector)
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_issues.yaml.
labels <- get_labels(selector = local_selector)
#> The labels will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_labels.yaml.
milestones <- get_milestones(selector = local_selector)
#> The milestones will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_milestones.yaml.