Get Issue Resolution Times

Calculates the time taken to resolve issues in seconds.

get_resolution_times(x, ...)

# S3 method for class 'IssuesTB'
get_resolution_times(x, verbose = TRUE, ...)

# Default S3 method
get_resolution_times(...)

Arguments

x

An object of class IssuesTB.

...

Currently not used.

verbose

A logical value indicating whether to print additional information. Default is TRUE.

Value

Integer vector of resolution times in seconds.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "closed_issues.yaml"
)
#> Looking into closed_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/closed_issues.yaml.

IssueTrackeR:::get_resolution_times(all_issues)
#>   [1]        0        0    86400        0        0    86400  6393600        0
#>   [9]        0  4147200  1987200  3801600        0  3024000  3024000  3024000
#>  [17]        0    86400  2073600    86400   345600        0 11923200        0
#>  [25]    86400 19958400  2332800    86400  2332800    86400  2851200    86400
#>  [33]   345600  6652800 10627200    86400 11577600 13305600  1555200        0
#>  [41]  1296000    86400  1468800  1987200  1987200        0  2592000  2419200
#>  [49] 29289600   259200  3801600    86400  4406400 18230400        0  4320000
#>  [57] 18316800  5356800  6048000        0        0  6825600  6739200        0
#>  [65]   259200   432000 29548800 15379200        0   172800  1468800  1987200
#>  [73]  3456000  3369600        0        0   259200  3456000  5529600  9590400
#>  [81] 19353600    86400  9936000  1728000    86400    86400    86400        0
#>  [89] 13046400 13478400 13478400 10108800 37670400    86400  1209600 14860800
#>  [97] 50716800   345600    86400 54432000        0 55036800   172800        0
#> [105]        0        0    86400        0    86400  1382400  1123200 33782400
#> [113] 27907200  1296000  1900800  2592000  2073600   518400  3801600  3974400
#> [121] 60048000  5097600  5097600  8467200    86400   172800        0  9158400
#> [129]        0   259200  9936000  1296000  1296000  2851200  3715200 12787200
#> [137] 15811200 16243200 76032000 71366400 51148800   259200   172800 17712000
#> [145]  9763200 18144000  5875200        0 28252800        0    86400 28771200
#> [153] 31017600   345600