Display IssueTB and IssuesTB object

Display IssueTB and IssuesTB with formatted output in the console

# S3 method for class 'IssueTB'
print(x, ...)

# S3 method for class 'IssuesTB'
print(x, ...)

# S3 method for class 'summary.IssueTB'
print(x, ...)

# S3 method for class 'summary.IssuesTB'
print(x, ...)

# S3 method for class 'LabelsTB'
print(x, ...)

# S3 method for class 'summary.LabelsTB'
print(x, ...)

Arguments

x

a IssueTB or IssuesTB object.

...

Unused argument

Value

invisibly (with invisible()) NULL.

Details

This function displays an issue (IssueTB object) or a list of issues (IssuesTB object) with a formatted output.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.

# Display one issue
print(all_issues[1, ])
#> 
#> ── Issue jdemetra/jdplus-main#844 ──
#> 
#> Title: Charts are slow when displaying lots of observations
#> Text:
#> High frequencies provide a lot of observations. 
#> More than 600000 observations seems to be problematic in charts.
#> (see `sdmx-dl:/STATCAN_HFED/CCEI%2CDF_HFED_AB%2C1.0/H.CA_AB.CA_BC.NSI`).
#> 

# Display several issues
print(all_issues[1:10, ])
#> There are 10 issues. 
#> 
#> ── Issue jdemetra/jdplus-main#844 ──
#> 
#> Title: Charts are slow when displaying lots of observations
#> Text:
#> High frequencies provide a lot of observations. 
#> More than 600000 observations seems to be problematic in charts.
#> (see `sdmx-dl:/STATCAN_HFED/CCEI%2CDF_HFED_AB%2C1.0/H.CA_AB.CA_BC.NSI`).
#> 
#> 
#> ── Issue jdemetra/jdplus-main#843 ──
#> 
#> Title: Why the color is not updated?
#> Text:
#> ## Questions
#> 
#> ### Not the color
#> 
#> When I try to switch from a Tramo-Seats specification to X13 (or vice versa), the `domainSpec` (**Reference Specification**) name is updated, but **not the color**.
#> 
#> ### Apply before refresh
#> 
#> Then, if I want to refresh the model, I have to **apply the `domainSpec`** before I can use a r
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#842 ──
#> 
#> Title: Zoom in GUI
#> Text:
#> It would be great to be able to zoom in the GUI.
#> 
#> For example, here is a normal view with big screen:
#> 
#> <img width="1920" height="1032" alt="Image" src="https://github.com/user-attachments/assets/e9f0a942-daa0-41b1-8980-09823cda08a9" />
#> 
#> We could zoom in = increase the size of the fonts, the width, the height of the con
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#841 ──
#> 
#> Title: Overview of a theme
#> Text:
#> ## Questions
#> 
#> ### Change and preview a theme
#> 
#> When you want to change the theme in the GUI, you have to restart the interface. It would be interesting to have a preview (without necessarily updating the interface) to choose a theme.
#> 
#> ### Light and dark
#> 
#> Also, in many programs, it is possible to choose a “light” and a “
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#840 ──
#> 
#> Title: White on white headers
#> Text:
#> In grid mode, headers turn white on white as you move the mouse over them.
#> 
#> https://github.com/user-attachments/assets/b6787118-3f06-4dcf-94bc-42fe4607fb4f
#> 
#> ── Issue jdemetra/jdplus-main#838 ──
#> 
#> Title: Cannot delete several objects at the same time
#> Text:
#> For certain types of object, it is not possible to delete them all at the same time.
#> 
#> These include:
#> 
#> - Several groups of variables:
#> 
#> <img width="310" height="226" alt="Image" src="https://github.com/user-attachments/assets/63bc2edf-a114-4a60-a230-485f7fb92808" />
#> 
#> - Several SA-Processings:
#> 
#> <img width="313" height="14
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#837 ──
#> 
#> Title: `Delete`, `Close`, `Clear` or `Remove`
#> Text:
#> The vocabulary used to remove or delete an element is not the same:
#> 
#> - **Remove** is used for variables
#> - **Delete** is used for SA-objects (SAI, SAP, SAD and WS)
#> - **Close** is used for provider connections
#> - **Clear** for containers (grid, group of regressors and chart)
#> 
#> 
#> ## Examples
#> 
#> ### `Remove`
#> 
#> Removing variables
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#821 ──
#> 
#> Title: Accepting new providers
#> Text:
#> With the emergence of new file formats, would it be costly to add new providers?
#> 
#> At Insee, we're thinking in particular of the parquet format (which is not currently a priority for time series data, but could become one in the future). 
#> 
#> ── Issue jdemetra/jdplus-main#820 ──
#> 
#> Title: Problem with creating csv and csv matrix output
#> Text:
#> When outputting the results of seasonal adjustment in csv and csv matrix format, the output fails with error `java.nio.charset.UnmappableCharacterException: Input Length = 1` (more details below). The problem was first seen on version 3.5.1 and is still seen in the newest version 3.6.0 using a computer running Windows 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#819 ──
#> 
#> Title: Regarima residuals threshold values
#> Text:
#> We have noticed that threshold values in `Tools > Options > SA > TramoSeats > Regarima residuals` are the same for Uncertain/Bad for Seasonal and Severe/Bad for Trading days (see picture below).
#> When changing the settings and closing the program, the changes are not saved, they default to the same value.
#> 
#> <img width="4
#> ...
#> 

# Display the summary of one issue
summary(all_issues[2, ])
#> ── Issue jdemetra/jdplus-main#843 ──
#> 
#> State: 🟢 Open
#> Nb comments: 0
#> 
#> Title: Why the color is not updated?
#> Text:
#> ## Questions
#> 
#> ### Not the color
#> 
#> When I try to switch from a Tramo-Seats specification to X13 (or vice versa), the `domainSpec` (**Reference Specification**) name is updated, but **not the color**.
#> 
#> ### Apply before refresh
#> 
#> Then, if I want to refresh the model, I have to **apply the `domainSpec`** before I can use a refresh policy (e.g. Concurrent). I'm not sure if this was the case with other specification changes or other refresh policies.
#> 
#> ### Refresh only the first
#> 
#> When I refresh a group of SA-Items and one of the refreshes fails, only the first n are executed (even if there are valid ones that haven't been tested).
#> In my example, only the first SA-Item is refreshed.
#> 
#> ### **Bold** or `(frozen)`
#> 
#> Finally, the name of an SA-Item can be **bold* or followed by `(frozen)`. What does this mean again?
#> 
#> I think it means it's linked to fact that the link to the raw data is updated or that the the model is refreshed ?
#> 
#> There are cases where the `(frozen)` remains in the SA-Item name, I need to raise an issue for that too.
#> 
#> 
#> ## Examples
#> 
#> Here is an example:
#> 
#> https://github.com/user-attachments/assets/9f6e1e62-2aaa-40c3-8a8d-5313c06c6bfa
#> 
#> ### Errors
#> 
#> The error message I got:
#> 
#> https://github.com/user-attachments/assets/3273b9af-d573-440d-aec3-ea5581e50fa1
#> 
#> 
#> <details>
#>   <summary>Tramo-Seats full log</summary>
#> 
#> 
#> ```
#> java.lang.ClassCastException: class jdplus.x13.base.api.x13.X13Spec cannot be cast to class jdplus.tramoseats.base.api.tramoseats.TramoSeatsSpec (jdplus.x13.base.api.x13.X13Spec is in unnamed module of loader org.netbeans.StandardModule$OneModuleClassLoader @7cf564fc; jdplus.tramoseats.base.api.tramoseats.TramoSeatsSpec is in unnamed module of loader org.netbeans.StandardModule$OneModuleClassLoader @30e1fd5a)
#> 	at jdplus.tramoseats.base.core.tramoseats.TramoSeatsFactory.refreshSpec(TramoSeatsFactory.java:62)
#> 	at jdplus.sa.base.api.SaItem.refresh(SaItem.java:311)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.ui.MultiProcessingDocument.refresh(MultiProcessingDocument.java:174)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.ui.SaBatchUI.refresh(SaBatchUI.java:508)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.ui.SaBatchUI.refresh(SaBatchUI.java:493)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.actions.LocalRefreshConcurrent.process(LocalRefreshConcurrent.java:37)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.actions.LocalRefreshConcurrent.process(LocalRefreshConcurrent.java:18)
#> 	at jdplus.toolkit.desktop.plugin.ui.ActiveViewAction.actionPerformed(ActiveViewAction.java:42)
#> 	at java.desktop/javax.swing.AbstractButton.fireActionPerformed(Unknown Source)
#> 	at java.desktop/javax.swing.AbstractButton$Handler.actionPerformed(Unknown Source)
#> 	at java.desktop/javax.swing.DefaultButtonModel.fireActionPerformed(Unknown Source)
#> 	at java.desktop/javax.swing.DefaultButtonModel.setPressed(Unknown Source)
#> 	at java.desktop/javax.swing.AbstractButton.doClick(Unknown Source)
#> 	at java.desktop/javax.swing.plaf.basic.BasicMenuItemUI.doClick(Unknown Source)
#> 	at java.desktop/javax.swing.plaf.basic.BasicMenuItemUI$Handler.mouseReleased(Unknown Source)
#> 	at java.desktop/java.awt.Component.processMouseEvent(Unknown Source)
#> 	at java.desktop/javax.swing.JComponent.processMouseEvent(Unknown Source)
#> 	at java.desktop/java.awt.Component.processEvent(Unknown Source)
#> 	at java.desktop/java.awt.Container.processEvent(Unknown Source)
#> 	at java.desktop/java.awt.Component.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Container.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Component.dispatchEvent(Unknown Source)
#> 	at java.desktop/java.awt.LightweightDispatcher.retargetMouseEvent(Unknown Source)
#> 	at java.desktop/java.awt.LightweightDispatcher.processMouseEvent(Unknown Source)
#> 	at java.desktop/java.awt.LightweightDispatcher.dispatchEvent(Unknown Source)
#> 	at java.desktop/java.awt.Container.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Window.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Component.dispatchEvent(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$4.run(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$4.run(Unknown Source)
#> 	at java.base/java.security.AccessController.doPrivileged(Unknown Source)
#> 	at java.base/java.security.ProtectionDomain$JavaSecurityAccessImpl.doIntersectionPrivilege(Unknown Source)
#> 	at java.base/java.security.ProtectionDomain$JavaSecurityAccessImpl.doIntersectionPrivilege(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$5.run(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$5.run(Unknown Source)
#> 	at java.base/java.security.AccessController.doPrivileged(Unknown Source)
#> 	at java.base/java.security.ProtectionDomain$JavaSecurityAccessImpl.doIntersectionPrivilege(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue.dispatchEvent(Unknown Source)
#> [catch] at org.netbeans.core.TimableEventQueue.dispatchEvent(TimableEventQueue.java:136)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpOneEventForFilters(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEventsForFilter(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEventsForHierarchy(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEvents(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEvents(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.run(Unknown Source)
#> ```
#> 
#> </details>
#> 
#> 
#> <details>
#>   <summary>X13 full log</summary>
#> 
#> 
#> ```
#> java.lang.ClassCastException: class jdplus.tramoseats.base.api.tramoseats.TramoSeatsSpec cannot be cast to class jdplus.x13.base.api.x13.X13Spec (jdplus.tramoseats.base.api.tramoseats.TramoSeatsSpec is in unnamed module of loader org.netbeans.StandardModule$OneModuleClassLoader @30e1fd5a; jdplus.x13.base.api.x13.X13Spec is in unnamed module of loader org.netbeans.StandardModule$OneModuleClassLoader @7cf564fc)
#> 	at jdplus.x13.base.core.x13.X13Factory.refreshSpec(X13Factory.java:70)
#> 	at jdplus.sa.base.api.SaItem.refresh(SaItem.java:311)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.ui.MultiProcessingDocument.refresh(MultiProcessingDocument.java:174)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.ui.SaBatchUI.refresh(SaBatchUI.java:508)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.ui.SaBatchUI.refresh(SaBatchUI.java:493)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.actions.LocalRefreshConcurrent.process(LocalRefreshConcurrent.java:37)
#> 	at jdplus.sa.desktop.plugin.multiprocessing.actions.LocalRefreshConcurrent.process(LocalRefreshConcurrent.java:18)
#> 	at jdplus.toolkit.desktop.plugin.ui.ActiveViewAction.actionPerformed(ActiveViewAction.java:42)
#> 	at java.desktop/javax.swing.AbstractButton.fireActionPerformed(Unknown Source)
#> 	at java.desktop/javax.swing.AbstractButton$Handler.actionPerformed(Unknown Source)
#> 	at java.desktop/javax.swing.DefaultButtonModel.fireActionPerformed(Unknown Source)
#> 	at java.desktop/javax.swing.DefaultButtonModel.setPressed(Unknown Source)
#> 	at java.desktop/javax.swing.AbstractButton.doClick(Unknown Source)
#> 	at java.desktop/javax.swing.plaf.basic.BasicMenuItemUI.doClick(Unknown Source)
#> 	at java.desktop/javax.swing.plaf.basic.BasicMenuItemUI$Handler.mouseReleased(Unknown Source)
#> 	at java.desktop/java.awt.Component.processMouseEvent(Unknown Source)
#> 	at java.desktop/javax.swing.JComponent.processMouseEvent(Unknown Source)
#> 	at java.desktop/java.awt.Component.processEvent(Unknown Source)
#> 	at java.desktop/java.awt.Container.processEvent(Unknown Source)
#> 	at java.desktop/java.awt.Component.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Container.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Component.dispatchEvent(Unknown Source)
#> 	at java.desktop/java.awt.LightweightDispatcher.retargetMouseEvent(Unknown Source)
#> 	at java.desktop/java.awt.LightweightDispatcher.processMouseEvent(Unknown Source)
#> 	at java.desktop/java.awt.LightweightDispatcher.dispatchEvent(Unknown Source)
#> 	at java.desktop/java.awt.Container.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Window.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.Component.dispatchEvent(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue.dispatchEventImpl(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$4.run(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$4.run(Unknown Source)
#> 	at java.base/java.security.AccessController.doPrivileged(Unknown Source)
#> 	at java.base/java.security.ProtectionDomain$JavaSecurityAccessImpl.doIntersectionPrivilege(Unknown Source)
#> 	at java.base/java.security.ProtectionDomain$JavaSecurityAccessImpl.doIntersectionPrivilege(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$5.run(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue$5.run(Unknown Source)
#> 	at java.base/java.security.AccessController.doPrivileged(Unknown Source)
#> 	at java.base/java.security.ProtectionDomain$JavaSecurityAccessImpl.doIntersectionPrivilege(Unknown Source)
#> 	at java.desktop/java.awt.EventQueue.dispatchEvent(Unknown Source)
#> [catch] at org.netbeans.core.TimableEventQueue.dispatchEvent(TimableEventQueue.java:136)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpOneEventForFilters(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEventsForFilter(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEventsForHierarchy(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEvents(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.pumpEvents(Unknown Source)
#> 	at java.desktop/java.awt.EventDispatchThread.run(Unknown Source)
#> ```
#> 
#> </details>
#> 

# Display the summary of
summary(all_issues[1:10, ])
#> There are 10 issues. 
#> - jdemetra/jdplus-main#844 🟢 Open 
#> - jdemetra/jdplus-main#843 🟢 Open 
#> - jdemetra/jdplus-main#842 🟢 Open 
#> - jdemetra/jdplus-main#841 🟢 Open 
#> - jdemetra/jdplus-main#840 🟢 Open 
#> - jdemetra/jdplus-main#838 🟢 Open 
#> - jdemetra/jdplus-main#837 🟢 Open 
#> - jdemetra/jdplus-main#821 🟢 Open 
#> - jdemetra/jdplus-main#820 🟢 Open 
#> - jdemetra/jdplus-main#819 🟢 Open