Sorting issues

Sort issues based on quantitative variables

simple_sort(issues, sorting_variables, milestones, ...)

Arguments

issues

a IssuesTB object.

sorting_variables

a list of variable (represented by couples of object and field in the form c(object = "object_1", field = "field_1"))

milestones

a list representing milestones with simpler structure (with title, description and due_on).

...

Additional arguments passed to get_milestones. Only used if milestones is missing.

Value

the same list of issues as issues (class IssuesTB) but in a sorted order.

Examples

all_issues <- get_issues()
#> The issues will be read from /tmp/RtmpIZvBFR/data/open_issues.yaml.
all_milestones <- get_milestones()
#> The milestones will be read from /tmp/RtmpIZvBFR/data/list_milestones.yaml.

# with milestones
IssueTrackeR:::simple_sort(
    issues = all_issues,
    milestones = all_milestones,
    sorting_variables = list(
        c(object = "milestones", field = "due_on"),
        c(object = "issues", field = "created_at")
    )
)
#> There are 4 issues. 
#> 
#> Issue #76
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Leap Year Effect 
#> Text:
#> I am running into problems including the leap year:
#> https://github.com/statistikat/persephone/issues/21
#> 
#> Am I missing some additional parameters that need to be set?  
#> 
#> 
#> Issue #142
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Printing userdef variables with their name 
#> Text:
#> When you display a model (obtained with the get_model function), the regression variables are not displayed with their real name but with a name given by jdemetra+:
#> 
#> ![image](https://github.com/jdemetra/rjdemetra/assets/112711017/2dc551f8-2c0a-4da0-bc12-d74285537247)
#> 
#> However, their variable name does exist in the object:
#> 
#> ![image](https://github.com/jdemetra/rjdemetra/assets/112711017/7645a392-6e06-4480-b832-282b83ec77da)
#> 
#> Is it possible to modify the display to add these more meaningful variable names instead of the current names? 
#> 
#> Comment:
#> They do not come from the same object and there are some differences in the Java output. 
#> In the first case the names comes from the results of the regarima model (<https://github.com/rjdverse/rjdemetra/blob/develop/R/regarima_rslts.R#L52>) and in the second case it comes from the specification (<https://github.com/rjdverse/rjdemetra/blob/develop/R/spec_rjd.R#L193-L220>). I think it would be complicated to do it in R since you would need to indentify your external regressors from all the regressions coefficients (outliers, easter, td, etc.), but maybe something can be done using the `context_dictionary`. Do you have a reproducible example? 
#> 
#> 
#> Issue #156
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Possible to extract slidingspan main diagnostics ? 
#> Text:
#> I wondered if it was possible in rjdemetra to extract and save the main slidingspan diagnostics of % unstable observations (A% and MM%) that are calculated  and displayed (but not saved) in JDemetra+ ?
#> 
#> I guess this is not an “issue”, but rather a (Christmas) wish
#> 
#> Best regards 
#> Jørn Ivar Hamre
#> (Statistics Norway)
#>  
#> 
#> Comment:
#> Hi,
#> 
#> it is indeed a wish ! We'll see...
#> 
#> Merry Xmas 
#> 
#> 
#> Issue #162
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: How to differentiate a WS in V2 and in V3? 
#> Text:
#> *related to rjdverse/rjd3workspace#53*
#> 
#> When I try to load a WS in v3 with **{RJDemetra}** (with `load_workspace()`), I get no error but I will get error when trying to access other elements.
#> 
#> How to know if a WS is in V3 with **{RJDemetra}**? Is it possible?  
#> 

# without milestones
IssueTrackeR:::simple_sort(
    issues = all_issues,
    sorting_variables = list(
        c(object = "milestones", field = "due_on"),
        c(object = "issues", field = "created_at")
    ),
    source = "local"
)
#> The milestones will be read from /tmp/RtmpIZvBFR/data/list_milestones.yaml.
#> There are 4 issues. 
#> 
#> Issue #76
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Leap Year Effect 
#> Text:
#> I am running into problems including the leap year:
#> https://github.com/statistikat/persephone/issues/21
#> 
#> Am I missing some additional parameters that need to be set?  
#> 
#> 
#> Issue #142
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Printing userdef variables with their name 
#> Text:
#> When you display a model (obtained with the get_model function), the regression variables are not displayed with their real name but with a name given by jdemetra+:
#> 
#> ![image](https://github.com/jdemetra/rjdemetra/assets/112711017/2dc551f8-2c0a-4da0-bc12-d74285537247)
#> 
#> However, their variable name does exist in the object:
#> 
#> ![image](https://github.com/jdemetra/rjdemetra/assets/112711017/7645a392-6e06-4480-b832-282b83ec77da)
#> 
#> Is it possible to modify the display to add these more meaningful variable names instead of the current names? 
#> 
#> Comment:
#> They do not come from the same object and there are some differences in the Java output. 
#> In the first case the names comes from the results of the regarima model (<https://github.com/rjdverse/rjdemetra/blob/develop/R/regarima_rslts.R#L52>) and in the second case it comes from the specification (<https://github.com/rjdverse/rjdemetra/blob/develop/R/spec_rjd.R#L193-L220>). I think it would be complicated to do it in R since you would need to indentify your external regressors from all the regressions coefficients (outliers, easter, td, etc.), but maybe something can be done using the `context_dictionary`. Do you have a reproducible example? 
#> 
#> 
#> Issue #156
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Possible to extract slidingspan main diagnostics ? 
#> Text:
#> I wondered if it was possible in rjdemetra to extract and save the main slidingspan diagnostics of % unstable observations (A% and MM%) that are calculated  and displayed (but not saved) in JDemetra+ ?
#> 
#> I guess this is not an “issue”, but rather a (Christmas) wish
#> 
#> Best regards 
#> Jørn Ivar Hamre
#> (Statistics Norway)
#>  
#> 
#> Comment:
#> Hi,
#> 
#> it is indeed a wish ! We'll see...
#> 
#> Merry Xmas 
#> 
#> 
#> Issue #162
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: How to differentiate a WS in V2 and in V3? 
#> Text:
#> *related to rjdverse/rjd3workspace#53*
#> 
#> When I try to load a WS in v3 with **{RJDemetra}** (with `load_workspace()`), I get no error but I will get error when trying to access other elements.
#> 
#> How to know if a WS is in V3 with **{RJDemetra}**? Is it possible?  
#>