Sort issues

Sorting issues with some constraint and order on the labels, the title, the milestones and/or the body.

# S3 method for class 'IssuesTB'
sort(
  x,
  decreasing = FALSE,
  sorting_variables = list(),
  filtering_factors = list(),
  ...
)

Arguments

x

a IssuesTB object.

decreasing

logical. Should the sort be increasing or decreasing?

sorting_variables

a list containing the quantitative variables to sort the issues. The filters are applied in the order of the variables supplied.

filtering_factors

a list containing constraints for sorting issues by sub-group in order of priority

...

Additional arguments related to milestones for the function simple_sort.

Value

a IssuesTB object sorted.

Details

In the order of the constraints imposed by the filtering_factors argument, the function will first filter by constraint. For each constraint, the function will then sort according to the quantitative variables supplied in sorting_variables.

For example, the following call:

sort(
    x = issues,
    sorting_variables = list(c(object = "milestones", field = "due_on"),
                             c(object = "issues", field = "created_at")),
    filtering_factors = list(list(values = "bug",
                                  fields = "labels",
                                  values_logic_gate = "OR"),
                             list(values = "package", fields = "title")),
    decreasing = TRUE
)

will behave as follows: 1) It will select all the issues that have "bug" as a label, then sort them according to the chronological order of milestones (according to deadlines) and the chronological order of issue creation dates 2) Among the remaining issues, it will filter the issues that have "package" in the title and apply the same sorting. 3) Finally, among all the remaining issues (not sorted until now), the function will apply the same sorting. 4) The function returns the global list of sorted issues.

The argument filtering_factors is a list of constraint following the same naming convention as the filter_issues. So the constraints are represented by named lists with the various arguments (apart from x) to the filter_issues (values, fields, fields_logic_gate, values_logic_gate and negate).

Examples

# Get the milestones of the prject
milestones <- get_milestones("online")
#> Reading milestones... 
#> Done! 0 milestones found.
write_milestones_to_dataset(milestones)
#> The datasets will be exported to /tmp/Rtmp0eoUsH/data/list_milestones.yaml.
#> The file already exists and will be overwritten.

all_issues <- get_issues(source = "online", verbose = FALSE)
sort(
    x = all_issues,
    sorting_variables = list(list(object = "milestones", field = "due_on"),
                             list(object = "issues", field = "created_at")),
    filtering_factors = list(list(values = "bug",
                                  fields = "labels",
                                  values_logic_gate = "OR"),
                             list(values = "package", fields = "title")),
    milestones = milestones
)
#> There are 4 issues. 
#> 
#> Issue #76
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Leap Year Effect 
#> Text:
#> I am running into problems including the leap year:
#> https://github.com/statistikat/persephone/issues/21
#> 
#> Am I missing some additional parameters that need to be set?  
#> 
#> 
#> Issue #142
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Printing userdef variables with their name 
#> Text:
#> When you display a model (obtained with the get_model function), the regression variables are not displayed with their real name but with a name given by jdemetra+:
#> 
#> ![image](https://github.com/jdemetra/rjdemetra/assets/112711017/2dc551f8-2c0a-4da0-bc12-d74285537247)
#> 
#> However, their variable name does exist in the object:
#> 
#> ![image](https://github.com/jdemetra/rjdemetra/assets/112711017/7645a392-6e06-4480-b832-282b83ec77da)
#> 
#> Is it possible to modify the display to add these more meaningful variable names instead of the current names? 
#> 
#> Comment:
#> They do not come from the same object and there are some differences in the Java output. 
#> In the first case the names comes from the results of the regarima model (<https://github.com/rjdverse/rjdemetra/blob/develop/R/regarima_rslts.R#L52>) and in the second case it comes from the specification (<https://github.com/rjdverse/rjdemetra/blob/develop/R/spec_rjd.R#L193-L220>). I think it would be complicated to do it in R since you would need to indentify your external regressors from all the regressions coefficients (outliers, easter, td, etc.), but maybe something can be done using the `context_dictionary`. Do you have a reproducible example? 
#> 
#> 
#> Issue #156
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: Possible to extract slidingspan main diagnostics ? 
#> Text:
#> I wondered if it was possible in rjdemetra to extract and save the main slidingspan diagnostics of % unstable observations (A% and MM%) that are calculated  and displayed (but not saved) in JDemetra+ ?
#> 
#> I guess this is not an “issue”, but rather a (Christmas) wish
#> 
#> Best regards 
#> Jørn Ivar Hamre
#> (Statistics Norway)
#>  
#> 
#> Comment:
#> Hi,
#> 
#> it is indeed a wish ! We'll see...
#> 
#> Merry Xmas 
#> 
#> 
#> Issue #162
#> rjdverse/rjdemetra
#> Labels:  
#> Milestone: 
#> Title: How to differentiate a WS in V2 and in V3? 
#> Text:
#> *related to rjdverse/rjd3workspace#53*
#> 
#> When I try to load a WS in v3 with **{RJDemetra}** (with `load_workspace()`), I get no error but I will get error when trying to access other elements.
#> 
#> How to know if a WS is in V3 with **{RJDemetra}**? Is it possible?  
#>