Update database

Update the different local database (issues, labels and milestones) with the online reference.

update_database(
  dataset_dir = getOption("IssueTrackeR.dataset.dir"),
  datasets_name = c(open = "open_issues.yaml", closed = "closed_issues.yaml", labels =
    "list_labels.yaml", milestones = "list_milestones.yaml"),
  verbose = TRUE,
  ...
)

Arguments

dataset_dir

A character string specifying the path which contains the datasets (only taken into account if source is set to "local"). Defaults to the package option IssueTrackeR.dataset.dir.

datasets_name

A named character string of length 4, specifying the names of the different datasets which will be written. The names datasets_name have to be "open", "closed", "labels" and "milestones". Defaults to c(open = "open_issues.yaml", closed = "closed_issues.yaml", labels = "list_labels.yaml", milestones = "list_milestones.yaml") .

verbose

A logical value indicating whether to print additional information. Default is TRUE.

...

Additional arguments for connecting to the GitHub repository:

repo A character string specifying the GitHub repository name. Defaults to the package option IssueTrackeR.repo.

owner A character string specifying the GitHub owner. Defaults to the package option IssueTrackeR.owner. (See the documentation of get to have more information on theses parameters):

Value

invisibly (with invisible()) TRUE.

Examples

# \donttest{
update_database(dataset_dir = tempdir())
#> Repo: rjdemetra  owner: rjdverse 
#> The datasets will be exported to /tmp/RtmpZS0YRs/open_issues.yaml.
#> Repo: rjdemetra  owner: rjdverse 
#> The datasets will be exported to /tmp/RtmpZS0YRs/closed_issues.yaml.
#> Repo: rjdemetra  owner: rjdverse 
#> Reading labels... Done!
#> 8 labels found.
#> The datasets will be exported to /tmp/RtmpZS0YRs/list_labels.yaml.
#> Repo: rjdemetra  owner: rjdverse 
#> Reading milestones... 
#> Done! 0 milestones found.
#> The datasets will be exported to /tmp/RtmpZS0YRs/list_milestones.yaml.
# }