Check for comments in GitHub Issues

Function to filter issues with (or without) comments.

with_comments(x, negate = FALSE)

# S3 method for class 'IssuesTB'
with_comments(x, negate = FALSE)

Arguments

x

An object of class IssuesTB.

negate

boolean indicating if we are searching for issues WITHOUT comments. Default is FALSE.

Value

An object IssuesTB with issues that satisfy the condition.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.
with_comments(all_issues)
#> There are 46 issues. 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#761 ──
#> 
#> Title: Reloading an xlsx files in Providers create new SA-Item
#> Text:
#> We observe what appears to be a bug.
#> 
#> ### Description
#> 
#> When reloading a data file in the Providers tab, new series are added to the open SA-Processing.
#> 
#> ### Procedure
#> 
#> 1) Open JDemetra+ v3.5.1
#> 2) Create a new WS
#> 3) Create a new SAP
#> 4) Add data from Providers to SAP
#> 5) Add new points to data file
#> 6) Re-add the data file
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#758 ──
#> 
#> Title: Unify period display in GUI (2)
#> Text:
#> Issue #489 is unresolved. 
#> The problem remains in version 3.5.1. 
#> 
#> Also, the SI ratio display is still different and appears with Q1, Q2, Q3, Q4 :
#> 
#> <img width="605" height="255" alt="Image" src="https://github.com/user-attachments/assets/49948a60-a027-4bed-8947-9e9eedc49537" />
#> 
#> ── Issue jdemetra/jdplus-main#751 ──
#> 
#> Title: Structure of a WS
#> Text:
#> Hello, 
#> 
#> I'm opening an issue but it's more of **a discussion**.
#> 
#> First of all, the WS structure contains:
#> 
#> - data (calendar, regressors and data), 
#> - statistical models (specifications) 
#> - and outputs. 
#> 
#> In the context of moving to a cloud-native solution and changing JD+ to a web API, would it make sense to review th
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#733 ──
#> 
#> Title: Changing the `domainSpec` doesn't refresh the model
#> Text:
#> ### Intro
#> 
#> In the latest v3.5.1, there is a *breaking change* with v3.5.0 regarding the specification handling.
#> 
#> ### Step to reproduce
#> 
#> 1) Open JDemetra+ v3.5.1
#> 2) Create a new WS
#> 3) Add a series (with for example RSA3 specification)
#> 4) Change the `domainSpec` (**Right click** > **Edit** > **Reference specification** >
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#695 ──
#> 
#> Title: Infinity result for diagnostics.fcast-outsample-variance causes remainder of dem
#> Text:
#> I've found a bug with the cruncher. 
#> 
#> Here's the output I saw at the end of my run: 
#> - At the end of the output printed out by cruncher was this: (note the 'failed' line)
#> ```
#> Flushing bundle...
#> Generate Output
#> Csv matrix
#> failed: Input length = 1
#> csv
#> generated
#> Saving new processing...
#> ```
#> - output series were produced a
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#690 ──
#> 
#> Title: Change of reference specification in SAProcessing is not working
#> Text:
#> I changed the reference specification from RSAfull to x11 and then the specification wasn't changed:
#> 
#> <img width="991" height="603" alt="Image" src="https://github.com/user-attachments/assets/719f1e55-d3f6-4c3e-9d54-6ac304e7a84e" />
#> 
#> ── Issue jdemetra/jdplus-main#664 ──
#> 
#> Title: Combined seasonality tests on sa not appearing
#> Text:
#> Hi,
#> 
#> When I set my (monthly) series and model span to only the last 36, 37 or 38 periods, the 'Combined seasonality tests on sa' and 'Combined seasonality tests on sa (last years)' don't appear in the Main results part of the GUI:
#> 
#> Main results with longer series and model spans:
#> ![Image](https://github.com/user-attach
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#653 ──
#> 
#> Title: Excel SA-Item names contain `\n` to separate sheet series
#> Text:
#> When we create a new SA-Item from an Excel file, the name of the series is automatically `sheet name\nseries name`.
#> 
#> On the GUI, it appears with an arrow ▶:
#> 
#> ![Image](https://github.com/user-attachments/assets/955bd1e0-98aa-4101-bc25-8af37974715b)
#> 
#> But on the XML file:
#> 
#> ![Image](https://github.com/user-attachments/asse
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#640 ──
#> 
#> Title: Role of the metadata in the WS structure
#> Text:
#> Currently in a SA-Processing there are SA-Item and each SA-Item can contain a `<item name="metadata">` section with (currently only) the `<item name="comment">` item.
#> I understand that the role of this section is to add information that are not related to the adjustment but we want to keep.
#> 
#> But could we include in thi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#636 ──
#> 
#> Title: Error with `mvn clean install -DskipTests`
#> Text:
#> I try to build the jdplus_main project with `mvn clean install -DskipTests`, but I get an error:
#> 
#> ```
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD FAILURE
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time:  07:58 min
#> [INFO] 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#601 ──
#> 
#> Title: What does 'Show all' do in right-click context menu of charts?
#> Text:
#> When I right-click on a chart and select 'Show all', it doesn't seem to do anything:
#> 
#> ![Image](https://github.com/user-attachments/assets/1a8f8989-f1b1-43f6-b6bf-0cf6d48babe7)
#> 
#> I tried highlighting and unhighlighting various series, and showing and hiding the legend and title, but 'Show all' doesn't seem to do anything
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#572 ──
#> 
#> Title: Cannot rename user-defined variables (regressors) successfully
#> Text:
#> Consider creating a Variables document with a regressor called 'x_1':
#> 
#> ![Image](https://github.com/user-attachments/assets/78726fbe-bfa3-4081-b8cb-43423b67eba2)
#> 
#> Now consider a series in a multi-document and applying the above regressor to it:
#> 
#> ![Image](https://github.com/user-attachments/assets/b39d481f-62e7-4f93-92fa
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#564 ──
#> 
#> Title: Missing item in dictionaries
#> Text:
#> I don't find some items in the dictionaries:
#> 
#> - in [RegressionDictionaries](https://github.com/jdemetra/jdplus-main/blob/develop/jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/main/java/jdplus/toolkit/base/api/dictionaries/RegressionDictionaries.java):
#>   - `regression.description`
#>   - `regressi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#562 ──
#> 
#> Title: Model changed by Seats - output
#> Text:
#> JDemetra 3.4.0 - When the TRAMO model is not decomposed, the representation (P,D,Q)(BP,BD,BQ)s  of the Seats model is not available.
#> 
#> ── Issue jdemetra/jdplus-main#557 ──
#> 
#> Title: Documentation for number pattern
#> Text:
#> Could it be possible to add some documentation to choose the **locale** corresponding to a `number pattern`:
#> 
#> ![Image](https://github.com/user-attachments/assets/0ae53d62-19d8-48a9-935f-37bdb174fde0)
#> 
#> Maybe a link or a small comment?
#> 
#> ── Issue jdemetra/jdplus-main#556 ──
#> 
#> Title: Add option to retrieve desktop application version in headless environment
#> Text:
#> Add option to retrieve desktop application version in headless environment.
#> 
#> Something like: 
#> ```shell
#> $ nbdemetra64.exe --nogui --nosplash --version
#> 
#> JDemetra+ 3.4.0
#> JVM: 21.0.1 (GraalVM Community OpenJDK 64-Bit Server VM 21.0.1+12-jvmci-23.1-b19)
#> OS: Windows 11 10.0 amd64
#> ```
#> 
#> ── Issue jdemetra/jdplus-main#534 ──
#> 
#> Title: Checking for invalid series in workspace files
#> Text:
#> When the JD+ GUI indicates that a series has a 'Status' of 'Invalid', how is this recorded in the underlying XML workspace file?
#> 
#> ![Image](https://github.com/user-attachments/assets/5e230a9f-148a-4a27-8a60-a83b3c654e90)
#> 
#> Is the 'Quality' tag changed to 'Undefined'? Is such a change sufficient, or is there some other in
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#532 ──
#> 
#> Title: Unnecessary row displayed in Grid of time series
#> Text:
#> The following series has been imported from an Excel spreadsheet via the Providers panel. The series spans April 1982 to December 1985. However, when viewing the series in a Grid by double clicking on it, a row for 1986 is unnecessarily shown:
#> 
#> ![Image](https://github.com/user-attachments/assets/368a03b5-4e6f-44e1-bfeb
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#523 ──
#> 
#> Title: Add Copy Path/Reference context menu item to the Providers tab
#> Text:
#> The SDMX plugin uses a Copy Path/Reference context menu item to allow the user to quickly copy and paste relevant data IDs such as source, flow, and key. It is also used to create a command line with the appropriate parameters, allowing for some automation.
#> 
#> ![Image](https://github.com/user-attachments/assets/63c3066f-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#492 ──
#> 
#> Title: V3 output dictionary for Cruncher parameter file
#> Text:
#> Regarding the output dictionary at: https://github.com/jdemetra/jwsacruncher/wiki/Output-dictionaries, this appears to be for JD+ version 2. If that is the case, could we please kindly request an exhaustive list for version 3?
#> 
#> Please also see the following separate but related issues (especially the first one):
#> - h
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#485 ──
#> 
#> Title: Series vs items in the Cruncher input parameter file
#> Text:
#> Regarding the output dictionary for the input parameter file for Cruncher (https://github.com/jdemetra/jwsacruncher/wiki/Output-dictionaries), could a column please be added to clarify which keys refer to series as opposed to items?
#> 
#> That is, which tags correspond to the `<matrix>` items block, and which tags corresp
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#483 ──
#> 
#> Title: Precedence of refresh policy specified in Cruncher parameter file vs workspace f
#> Text:
#> When supplying the input parameter file to a Cruncher run as documented at https://github.com/jdemetra/jwsacruncher/wiki#parameters-file, the refresh policy must be specified:
#> 
#> ![image](https://github.com/user-attachments/assets/7f0546af-a2f8-4b4a-8120-a538393f4f88)
#> 
#> However, a refresh policy can also be specified 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#458 ──
#> 
#> Title: Make release process more robust against missing binaries in java distribution
#> Text:
#> Some builds fail because the java distribution bundled in JD+ (temurin) has sometimes some issues with the windows binaries.
#> The release process should be more robust to handle this kind of event.
#> As an alternative, JD+ could also use Azul Zulu instead of Adoptium Temurin.
#> 
#> Current workaround: https://github.com/jd
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#455 ──
#> 
#> Title: Manual creation of user defined variables
#> Text:
#> Hey development team,
#> I am trying to directly create a workspace for JDemetra using its Java libraries (e.g. jdplus-sa-base-workspace). I want to add Variables and userDefinedVariables.
#> When creating the Variables using the JD GUI, the variables show up in the XML file as dynamicTsVariable. One can add multiple of th
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#428 ──
#> 
#> Title: Truncated text in global SA general options tab
#> Text:
#> Observe the truncated text in the General tab of the SA options window highlighted in the screen shot below. Despite sufficient room for the text to be displayed in full, it has been truncated. (Presumably, it should say 'Model stability'.)
#> 
#> ![image](https://github.com/user-attachments/assets/20dcec28-c637-4895-bd9b-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#423 ──
#> 
#> Title: Documentation on workspace file XML tags (seasonal adjustment parameter settings
#> Text:
#> Is there documentation on all possible XML tags (that is, all seasonal adjustment parameter settings as per the Specifications panel) that could possibly populate a workspace file using the X13 algorithm?
#> 
#> If not, could we please request such documentation, specifically doco that contains:
#> - all possible XML tags/pa
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#421 ──
#> 
#> Title: Missing revisions policy in documentation
#> Text:
#> According to the list of revisions policies for JDemetra+ at: https://github.com/jdemetra/jwsacruncher/wiki#parameters-file, there are 8 of them:
#> 
#> ![image](https://github.com/user-attachments/assets/807fdf98-3a8f-4fe0-b9fd-c5e3dd3d2eb7)
#> 
#> However, at: https://jdemetradocumentation.github.io/JDemetra-documentation/pa
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#399 ──
#> 
#> Title: How to change the 'bias' parameter in the GUI?
#> Text:
#> The XML file underlying each workspace includes the following `bias` tag by default:
#> 
#> ![image](https://github.com/user-attachments/assets/7d0bda1a-030e-4078-a06a-5041d2ca9c1f)
#> 
#> Two questions:
#> 
#> - What is it? (What does it mean or do, and what values can it take?)
#> - Where can it be viewed, changed or set in the Sp
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#385 ──
#> 
#> Title: Cannot save updates to user-defined variables (regressors) in the Variables docu
#> Text:
#> When I make changes to user-defined variables in the Variables document of my workspace, certain changes such as inserting new ones or removing existing ones, will not be saved. In one instance, when I click on 'Save Workspace', the underlying XML file is not updated, and in another instance, I cannot even select 'Save
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#381 ──
#> 
#> Title: Allow separation of excluding forecasts vs backcasts in the generation of extrem
#> Text:
#> A potential enhancement of the 'Excludeforecast' parameter is to separate it into excluding forecasts and excluding backcasts individually, so that forecasts can be excluded in the generation of extreme values whilst backcasts are not excluded, or vice versa.
#> 
#> ![image](https://github.com/user-attachments/assets/54657
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#372 ──
#> 
#> Title: The name of the newly created WS does not appear in the GUI
#> Text:
#> When a ws is created and saved in version 3, its name does not appear in the bar provided.
#> To make it appear, close the ws then reopen it and the name will be displayed correctly.
#> 
#> Here is a small demo of what I did to get this bug:
#> 
#> https://github.com/jdemetra/jdplus-main/assets/112711017/963115ec-a1ac-421f-b237-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#349 ──
#> 
#> Title: Icons to expand and collapse panes in multi-processing window too small
#> Text:
#> Is it intentional to have the buttons highlighted in the image below so small? They are very hard to click since they are so small, and I suspect many users miss seeing them because they are so small, especially in the latter image.
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/110582146/afccad96-d618-44bc
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#345 ──
#> 
#> Title: Format type of date in output
#> Text:
#> Dear all, 
#> 
#> I noticed some trouble with the new output format: 
#> in version 2, the date was of format "General" in the output, however 
#> in version 3, the date is changed to a real date of format type: 
#> 1995-01-01T00:00:00/PM3, which creates problems in our subsequent macros. 
#> Is there any special reason behind thi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#338 ──
#> 
#> Title: Restoring series name in legend does nothing
#> Text:
#> When I rename a series in the legend of a chart and click on 'Restore', nothing seems to happen. For example, in the image below, the initial names of the series that I used to create the chart were both 'Seasonally adjusted'. I changed this to 'SA (with auto-outliering)' and 'SA (without auto-outliering)'. When I then
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#335 ──
#> 
#> Title: Open JDemetra+ workspace with `Open with...`
#> Text:
#> When I want to open an existing workspace with JDemetra+, I open the GUI then I go to `File` --> `Open Workspace` --> then I search my `.xml` file.
#> 
#> Is it possible to open a workspace (from the `.xml` file) with the `Open with...` button?
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/112711017/d21266de-7
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#315 ──
#> 
#> Title: Missing diagnostics in Summary
#> Text:
#> In the main panel of a SA-Item (**Main Results**), there are sometimes some missing diagnostics in the **Residuals seasonality tests**.
#> 
#> A classic series contains 4 items in the **Residuals seasonality tests** : two QS-Test and two F-test:
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/112711017/9dafb92b-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#283 ──
#> 
#> Title: Plot and display series if SA fail (and additionnal explanations)
#> Text:
#> (Issue related to jdemetra/jdemetra-core#1129)
#> 
#> If the seasonnal adjustment fails, I can't neither plot nor see my values.
#> 
#> For example here, I have a dataset with 11 values so the seasonal adjustment will obviously fail:
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/112711017/818fc2b7-9885-4627-8edc-b
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#270 ──
#> 
#> Title: warning at start of 3.2.2
#> Text:
#> warning message when starting 3.2.2
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/43274870/e4291603-3036-4f16-a1b5-a42dd360af11)
#> 
#> Details 
#> org.netbeans.core.network.proxy.pac.PacParsingException: Cannot find secure PAC script engine.
#> Allowed engines: GraalVM:js,Graal.js,Nashorn
#> Found engines:
#> Will no
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#261 ──
#> 
#> Title: No new logo in shortcuts
#> Text:
#> the new logo appears just in the top left corner of the open GUI but not in the short cuts 
#> 
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/43274870/d5c4ca0c-9d70-42ee-bd40-97a2714db224)
#> 
#> 
#> ── Issue jdemetra/jdplus-main#250 ──
#> 
#> Title: Add time filter in UI components
#> Text:
#> See https://github.com/jdemetra/jdplus-main/blob/develop/jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/main/java/jdplus/toolkit/base/api/timeseries/TimeSelector.java
#> 
#> ── Issue jdemetra/jdplus-main#177 ──
#> 
#> Title: Improve branding
#> Text:
#> Current branding of JD+ is outdated and inconsistent.
#> 
#> Here is a non-exhaustive list of items to update and improve:
#> 1. icon, favicon and avatar (square shape, logo only)
#> 2. splash screen, about dialog image and banner (rectangular shape, landscape orientation, logo+text)
#> 3. short description (a few words)
#> 4. lon
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#163 ──
#> 
#> Title: Improve I/O of CSV files
#> Text:
#> The current import/export from/to CSV files has the following problems:
#> - The UI is inconsistent and confusing.
#> - The code is redundant and all over the place.
#> 
#> ── Issue jdemetra/jdplus-main#161 ──
#> 
#> Title: Use enhanced Pseudo-Random Number Generators from JDK 17
#> Text:
#> See https://openjdk.org/jeps/356
#> 
#> ── Issue jdemetra/jdplus-main#160 ──
#> 
#> Title: Migrate to Java 21
#> Text:
#> The current Java version used by JD+ is Java 17.
#> A new Java LTS has been released and brings significant improvements in performance, language and libraries.
#> 
#> See https://nipafx.dev/road-to-21-upgrade/
#> 
#> ── Issue jdemetra/jdplus-main#136 ──
#> 
#> Title: [FR] decimal separator in the JWSACruncher
#> Text:
#> It could be good to be able to change the decimal separator in the JWSACruncher to avoid a system dependency. This is linked to https://github.com/jdemetra/jwsacruncher/issues/13.
#> 
#> ── Issue jdemetra/jdplus-main#112 ──
#> 
#> Title: JWSACruncher and parameters `fullseriesname` and `refreshall`
#> Text:
#> I tried the following parameters files an a workspace with data created by an Excel file and it has no effect. Are the parameters `fullseriesname` and `refreshall` suppose to work in the JWSACruncher? Am I doing something wrong?
#> ```
#> <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
#> <wsaConfig bundle="10000" cs
#> ...
#> 
with_comments(all_issues, negate = TRUE)
#> There are 51 issues. 
#> 
#> ── Issue jdemetra/jdplus-main#763 ──
#> 
#> Title: Error loading V2 workspace with fixed week calendar day
#> Text:
#> If I create a workspace in V2 with a National Calendar with a Fixed Week it cannot be imported with the V3. Error loading the calendar with the following error:
#> ```
#> java.lang.NullPointerException: Cannot invoke "String.toUpperCase(java.util.Locale)" because "day" is null
#> 	at internal.toolkit.base.workspace.file.xml.uti
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#752 ──
#> 
#> Title: Add ways to retrieve WS file names and paths
#> Text:
#> It would be very interesting to add ways to retrieve WS file names and paths.
#> 
#> Is it possible to add methods like `getSource` and `setSource` to be able to retrieve the file containing a SA-Processing, a SA-Item or even a Workspace directly from the Java object?
#> 
#> Thank you
#> 
#> ── Issue jdemetra/jdplus-main#734 ──
#> 
#> Title: Can't refresh a SA-Item when WS newly created
#> Text:
#> ### Introduction
#> 
#> In JDemetra+ v3, it's not possible to refresh a SA-Item with **Right click** > **Refresh** when the newly created WS is not saved.
#> It was possible to refresh a new WS in version 2.
#> 
#> But there is A way to refresh a SA-Item with concurrent policy with the GUI in version 3.5.0 with a newly created WS is 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#711 ──
#> 
#> Title: GUI unresponsive when an incorrectly specified file is opened in the Providers p
#> Text:
#> Consider importing a CSV file in the Providers panel and specifying an incorrect date format (such as using 'mm' rather than 'MM', where the former specifies minutes and the latter specifies the month):
#> 
#> <img width="462" height="501" alt="Image" src="https://github.com/user-attachments/assets/8fac7b3d-29cd-4078-81a3-18
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#698 ──
#> 
#> Title: Series naming restrictions
#> Text:
#> This is just a question:
#> 
#> Are there any character restrictions imposed by the JDemetra+ GUI and/or the Cruncher when naming series (that is, when providing a name or ID for series)? For example, are certain special characters not allowed, or are there any character-length limits, and do these differ for the GUI vs the 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#689 ──
#> 
#> Title: Drag and Drop a timeseries from the results to the SA-Processing doubles the ser
#> Text:
#> 1. Step run a seasonal Adjustment
#> 
#> <img width="1028" height="703" alt="Image" src="https://github.com/user-attachments/assets/259b11fd-dece-4ac6-b37b-b2a4a7335d6c" />
#> 
#> 2. Drag and Drop a series
#> 
#> <img width="988" height="364" alt="Image" src="https://github.com/user-attachments/assets/f21f58fe-90d7-44b0-98f0-91c20e3cb1b
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#687 ──
#> 
#> Title: Output from Cruncher and Output from UI have a small inconsistency for fixed use
#> Text:
#> Please find enclosed the example for a user-defined trading day. The output could be generated with the cruncher or with output in the UI for a SaProcessing. 
#> In the output in "ahaCN_GP14\SAProcessing-1\demetra_m.csv" for ahaCN_GP14, the regression.ntd is 1 but the column E and F for the variables are empty. In Version
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#682 ──
#> 
#> Title: Make outlier information consistent in Pre-processing panel for fixed and non-fi
#> Text:
#> All fixed and non-fixed (estimated) outlier corrections can be conveniently seen in the Pre-processing panel. However, the information displayed for each type (fixed vs non-fixed) differs. In the example screen shot below, non-fixed (pre-specified) outliers have the type and period shown, but not the name; whereas for 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#679 ──
#> 
#> Title: Cruncher automatically changing transform function
#> Text:
#> Hi,
#> 
#> We have noticed that when a series that has regarima.transform.function = 'Log' cannot be logged (for example, if it has 0s or negative values in the original series), the Cruncher automatically converts the transform function to 'None'.
#> 
#> Are you able to confirm that this is the only situation in which the Crunche
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#667 ──
#> 
#> Title: Cannot scroll a long list of pre-specified outliers
#> Text:
#> For series with many pre-specified outliers, the left-hand panel of the 'Pre-specified outliers' dialogue box does not allow scrolling to see the entire list:
#> 
#> ![Image](https://github.com/user-attachments/assets/b3e5ade3-02b4-4b05-8d45-9d694812433b)
#> 
#> Fortunately, any non-visible outliers can still be navigated to using
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#661 ──
#> 
#> Title: Make error in saving workspaces explicit
#> Text:
#> We would like to suggest the following feature request that will improve the user experience:
#> 
#> The JD+ GUI does not provide an explicit error when the operation to save a workspace file has failed. Instead, the operation fails mostly silently, which misleads the user into thinking that their workspace has saved success
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#652 ──
#> 
#> Title: The cruncher generates no warning if failing to compute
#> Text:
#> We can study 3 different cases of data:
#> 
#> - ts without metadata
#> - ts with metadata
#> - ts with false metadata
#> 
#> In these 3 cases, the seasonal adjustment works, but the refresh only works for the first 2 cases and generates an invalid for the last case.
#> 
#> **Is this a deliberate behaviour?**
#> 
#> here is my R code:
#> 
#> ```r
#> library
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#651 ──
#> 
#> Title: How do you re-order SAIs in SAP?
#> Text:
#> Is it possible to move (up or down) an SA-Item in an SA-processing?
#> 
#> I'd like to reorganise my SA-Items so that they follow a certain nomenclature (a certain order).
#> I'd like to do this in R but I haven't found any other method than creating a copy of my WS and transferring my SA-Items one by one in the order I want.
#> 
#> 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#645 ──
#> 
#> Title: `No information for this item` and `Invalid` when Variables have spaces
#> Text:
#> When you create a WS with variables and the variable name contains one or more spaces, JDemetra+ no longer knows how to distinguish variables from their name and you end up with more variables than you expected and incorrect variables.
#> 
#> Here's an example of my tests on video:
#> 
#> https://github.com/user-attachments/assets
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#643 ──
#> 
#> Title: Several synonyms and imprecisions in the JDemetra+ term
#> Text:
#> In JDemetra+, there is a precise vocabulary for distinguishing objects.
#> **Names are important** and some terms are unfortunately synonyms and can refer to the same object.
#> 
#> I suggest grouping the names together to obtain one name = one object.
#> 
#> Here's my list:
#> | Current name                                             
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#642 ──
#> 
#> Title: Error with `mvn clean install`
#> Text:
#> I try to build the jdplus_main project with `mvn clean install`, but I get an error:
#> 
#> ```
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD FAILURE
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time:  01:38 h
#> [INFO] Finished at: 2
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#617 ──
#> 
#> Title: Cannot use keyboard shortcut to paste series into an empty grid container
#> Text:
#> In JD+ v3.4.0, I can copy a series and paste it into a chart container using the `Ctrl` + `V` keyboard shortcut. This works whether the chart container is empty, or whether it contains series already.
#> 
#> For grid containers however, the keyboard shortcut does not seem to work unless the user activates an existing column 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#608 ──
#> 
#> Title: Add export to mermaid.js chart
#> Text:
#> Description from Mermaid.js :
#> > JavaScript based diagramming and charting tool that renders Markdown-inspired text definitions to create and modify diagrams dynamically.
#> 
#> Example :
#> 
#> This text ...
#> ```
#> xychart-beta
#>     title "Sales Revenue"
#>     x-axis [jan, feb, mar, apr, may, jun, jul, aug, sep, oct, nov, dec]
#>     y-axi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#607 ──
#> 
#> Title: Cannot select all in a grid
#> Text:
#> When I click on **Select all** from the right-click context menu of a grid (first image below), nothing happens (second image below):
#> 
#> ![Image](https://github.com/user-attachments/assets/ea086707-b78d-4bf3-872a-c272939b9c4a)
#> 
#> ![Image](https://github.com/user-attachments/assets/f8f359a2-76f7-472e-a3dd-5da8a5ed22e8)
#> 
#> How
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#603 ──
#> 
#> Title: Cannot edit format of charts
#> Text:
#> When I right-click on a chart container and select **Edit format...** to change the formatting of the dates and numbers (such as using the settings below), nothing changes:
#> 
#> ![Image](https://github.com/user-attachments/assets/60d7c8fb-8724-461c-9fe0-3664dbbf7da0)
#> 
#> ![Image](https://github.com/user-attachments/assets/315
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#596 ──
#> 
#> Title: Add OpenNodeAction as preferred action on ProviderNode when node has no children
#> Text:
#> A provider node has no children if no datasource has been defined and/or restored.
#> Double-clicking on it in this state doesn't do anything.
#> Currently, the user must right-click the provider node to add a new source.
#> The idea is to check if the provider node has no children on double-click and, if so, add a new source d
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#594 ──
#> 
#> Title: Missing series items in the output of the GUI
#> Text:
#> Related to #294 
#> 
#>  ### GUI v3 vs {rjd3x13}
#> 
#> Normally, the graphical interface and R packages in version 3 produce the same output. 
#> However, some items produce an output with **{rjd3x13}** that cannot be obtained via the GUI output:
#> 
#> - `decomposition.s_cmp_f`
#> - `decomposition.sa_cmp_f`
#> - `decomposition.t_cmp_f`
#> - `l`
#> -
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#593 ──
#> 
#> Title: Periodicity/frequency of series not immediately clear
#> Text:
#> For series in a multi-document, it's not immediately obvious to the user what the periodicity or frequency of the series is. After processing an individual series and clicking on its row, the 'Main results' tab appears and the user can tell the frequency by the x-axis labelling of the SI chart as highlighted in red bel
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#591 ──
#> 
#> Title: User-defined variables renaming is not consistent
#> Text:
#> When renaming User-defined variables through the following menu, I can sometimes rename variables to the same name, and other times I get an error message.
#> 
#> ![Image](https://github.com/user-attachments/assets/b2b30a7e-cc26-44c0-9b71-4fb3c7b8bb95)
#> 
#> Here, I attempted to rename Headline_impute_2qtrs to Headline imputation
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#590 ──
#> 
#> Title: Summary info on pre-specified outliers not clear
#> Text:
#> In the JD+ GUI, the user can insert 'pre-specified outliers' as shown below, and these can be fixed or non-fixed:
#> 
#> ![Image](https://github.com/user-attachments/assets/d4a65a9a-3173-4e67-80a3-1e1f4cdd7ee7)
#> 
#> If the user then looks at the **Summary** section of the **Main **results**** section in the output area for the s
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#571 ──
#> 
#> Title: List of user-defined variables (regressors) displayed poorly in 'Variables' dial
#> Text:
#> When I attempt to insert/modify/remove outliers via the 'Variables' dialogue box, the name of the variable that I am currently editing does not appear correctly. In the screen shot below, I have clicked the `+` button to insert a new variable, but it is barely visible on the left-hand side of the box. It seems to have 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#560 ──
#> 
#> Title: XML files are badly indented from a certain level upwards
#> Text:
#> Everything is in the title. From a certain level, the indentation of the XML files comes back to 0:
#> 
#> ![Image](https://github.com/user-attachments/assets/5eac083a-6ea7-4dcd-b271-47916001eb47)
#> 
#> This makes reading these files very complicated. Is it possible to indent these files correctly?
#> 
#> ── Issue jdemetra/jdplus-main#559 ──
#> 
#> Title: Default MCD in MCDFilter
#> Text:
#> When all the I/C ratios are greater than 1, the MCD value used in the program is equal to 2: 
#> https://github.com/jdemetra/jdplus-main/blob/37c578dbff38f6c54f682536222c185c718eb4d7/jdplus-main-base/jdplus-x13-base-parent/jdplus-x13-base-core/src/main/java/jdplus/x13/base/core/x11/filter/X11FilterFactory.java#L40-L60
#> Is 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#510 ──
#> 
#> Title: Series selection blocking highlighting of rows in check last test
#> Text:
#> Regarding the check last function in the JDemetra+ GUI, the red/orange highlighting in the 'Rel. Error N-1' column is blocked whenever that series' row is selected. In the example below, the series in the first row is selected, but the red highlighting in the relative error column cannot be seen:
#> 
#> ![image](https://gi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#491 ──
#> 
#> Title: Add confidence interval in the distribution chart of the residuals.
#> Text:
#> *Issue revealed by Nikolina Rizanovska*
#> 
#> Is it possible to add dashed lines to indicate the confidence interval at 95% in the distribution chart of the residuals?
#> (We are in the `Preprocessing` --> `Residuals` --> `Distribution` section)
#> 
#> ![image](https://github.com/user-attachments/assets/1ed8e0ef-bcf7-48cd-aac5-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#488 ──
#> 
#> Title: Should leap year be considered as trading day regressor?
#> Text:
#> *Issue revealed by Nikolina Rizanovska*
#> 
#> When you want to use calendar regressors (**including the leap year**) in your seasonal adjustment, JDemetra+ indicates a number of variable (here 6) in the Summary:
#> 
#> ![image](https://github.com/user-attachments/assets/e41846f1-dba3-4abe-8424-8a152a458e9f)
#> 
#> But it doesn't 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#484 ──
#> 
#> Title: Cruncher error logs
#> Text:
#> Hi JDemetra+ devs,
#> We are using cruncher v3 in a data processing application implemented in Java. At a high level we are calling jdplus.cruncher.App.main with a valid workspace and parameters file, then parsing the output files for consumption by downstream services.
#> One issue with cruncher is that in most cases it d
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#479 ──
#> 
#> Title: Possible mistake in documentation for unit root limit options
#> Text:
#> Regarding the entry for the unit root limit on this page: https://jdemetradocumentation.github.io/JDemetra-documentation/pages/reference-manual/modelling-spec-arima.html
#> 
#> ![image](https://github.com/user-attachments/assets/03153d3f-56d4-439a-83d5-fd363090fcb4)
#> 
#> In the underlying XML workspace file, the correspondin
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#449 ──
#> 
#> Title: demetra_m.csv empty column names
#> Text:
#> In the demetra_m.csv file, empty column names appear as in version 2. It would be more useful to have a standard name for each column.
#> ![demetram_empty_columnames](https://github.com/user-attachments/assets/f08575e5-ed8a-40df-8ca8-99c4af149dd1)
#> 
#> 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#447 ──
#> 
#> Title: user-defined calendar effect names do not appear in demetra-m.csv
#> Text:
#> user-defined calendar effect names do not appear in demetra-m.csv file.
#> ![calendar effec_demetram](https://github.com/user-attachments/assets/267217bf-e37c-40bd-913e-324a5dc48d79)
#> 
#> 
#> 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#429 ──
#> 
#> Title: Ability to specify non-standard parameters for prior corrections is absent in th
#> Text:
#> In RJDemetra, we are able to specify additional parameters for prior corrections as appropriate. For example, for level shifts (LSs), we can choose to have them as being 'zero-ended' or 'one-ended'. Is this possible in the GUI? There is no option in the relevant dialogue box for LSs, for example:
#> 
#> ![image](https://gi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#420 ──
#> 
#> Title: Adding a message when a WS is saved
#> Text:
#> Is it possible to add a message (for example in the bottom right-hand corner of the interface) when a worskpace is saved to inform the user that the WS has been saved at its place or an error if it isn't the case ?
#> 
#> ── Issue jdemetra/jdplus-main#419 ──
#> 
#> Title: User-defined calendar effects
#> Text:
#> Better handling of user-defined calendar effects (output, gui)
#> 
#> ── Issue jdemetra/jdplus-main#418 ──
#> 
#> Title: Better handling of coefficients/parameters in the GUI
#> Text:
#> Improve the naming of coefficients/parameters in the spec boxes
#> Hide coefficients when they can't be specified (undefined transformation)
#> 
#> ── Issue jdemetra/jdplus-main#400 ──
#> 
#> Title: How to specify the data format in Providers?
#> Text:
#> When I try to add some series (txt files) in the Providers, by default there is a data format `1,234.5` (representing the value 1234.5):
#> 
#> ![image](https://github.com/user-attachments/assets/77695364-1999-468a-a066-502e323fb45a)
#> 
#> But how to change it with other formats as `1.234,5`.
#> There are a lot of different for
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#371 ──
#> 
#> Title: Add message in the case of redundant regression variables
#> Text:
#> It seems to me that in the presence of multicollinearity JD+ will declare one regression variable from each possible linear combination as being redundant and automatically set the corresponding regression effects to zero. The latter are then reported in a separate output table without any further explanation.
#> I think
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#353 ──
#> 
#> Title: Restore support of ODBC connection string on files 
#> Text:
#> In v2, it was possible to drag&drop files to the ODBC provider in order to create a connection string on the fly.
#> It was limited to MS Access files but could be extended to any file type supported by an ODBC driver.
#> 
#> ![Screenshot 2024-06-27 084947](https://github.com/jdemetra/jdplus-main/assets/8778378/2d58f93c-07f3
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#337 ──
#> 
#> Title: Possible enhancement of the GUI
#> Text:
#> Dear all, 
#> I noticed that if an SA-Processing is renamed, the new name is immediately reflected in the SA-Processing window but not in the menu bar (only after closing and reopening the Workspace). Would it be possible to fix this please? 
#> 
#> Best greetings, 
#> Véronique
#> 
#> ── Issue jdemetra/jdplus-main#331 ──
#> 
#> Title: Improve SA output
#> Text:
#> Two problems with current implementation:
#> - the notification doesn't show the created file on Excel output
#> ![Screenshot 2024-06-14 133226](https://github.com/jdemetra/jdplus-main/assets/8778378/4664035b-69e5-4d18-a656-4f49a7104baa)
#> ![Screenshot 2024-06-14 133329](https://github.com/jdemetra/jdplus-main/assets/877837
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#296 ──
#> 
#> Title: Display jump discontinuity for level shifts (trend breaks)
#> Text:
#> It would be ideal to have the option available to view a jump discontinuity at any level shift (trend break) when plotting the trend. In the image below, for example, instead of the blue trend line continuing unbroken, there is a visual jump where the level shift occurs in Q4 2001. This makes is easier and quicker for 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#293 ──
#> 
#> Title: Add support of standard reporting period and time range
#> Text:
#> JD+ only supports dates (`2010-01-01`) as I/O for observation periods.
#> Other representations exists as well and could be integrated into a future release such as:
#> - standard reporting period (`2010-Q1`, `2012-M12`)
#> - time range (`2010-01-01/P3M`, `2012-12-01/P1M`)
#> 
#> These representations are already defined in the 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#287 ──
#> 
#> Title: Show internal corrections on SI chart
#> Text:
#> There are two potential improvements to SI charts in the JDemetra+ GUI:
#> 
#> - revealing X11 internal automatic corrections by simultaneously displaying modified and unmodified SI ratios with lines joining them
#> - (if not the above, or possibly in addition to:) allowing the user to choose between displaying modified or u
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#272 ──
#> 
#> Title: SAVING WS: no message when overwriting an existing WS 
#> Text:
#> when saving a WS no message when overwriting an existing ws 
#> 
#> would be useful
#> 
#> ── Issue jdemetra/jdplus-main#243 ──
#> 
#> Title: Color scheme selection via 'Configure...' not user-friendly
#> Text:
#> (Version: 3.2.2-SNAPSHOT)
#> 
#> - A graph's display options can be modified via the 'Configure...' context menu when right-clicking on it. This brings up a dialogue box with an option to change the 'Color scheme'. However, the different colour options are not presented as a user-friendly drop-down list. Instead, the user 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#162 ──
#> 
#> Title: Use Platform-Specific Desktop Features from JDK 9
#> Text:
#> See https://openjdk.org/jeps/272
#> 
#> Most [nbbrd desktop API](https://github.com/nbbrd/java-desktop-util/) can be replaced by the latest JDK desktop API. 
#> There are a few exceptions anyway: search and kwown folders
#> 
#> ── Issue jdemetra/jdplus-main#130 ──
#> 
#> Title: XSD avaliability
#> Text:
#> As our organization are moving towards XMLs, it is getting too time consuming to adopt the XML files to the structure that JDEMETRA+ requries for input. 
#> 
#> Have this very easy code snippet to do it, given you have pivoted it in a cetrain way, but should be covered by someone who has far better knowledge. 
#> 
#> For bette
#> ...
#>