Name of last commentator

Retrieve the name of the last commentator

author_last_comment(x)

# S3 method for class 'IssueTB'
author_last_comment(x)

# S3 method for class 'IssuesTB'
author_last_comment(x)

Arguments

x

An object of class IssueTB or IssuesTB.

Value

A string with the name of the last person which leaves a comment. If there is no comments, it returns an empty string.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.
author_last_comment(all_issues)
#>   [1] ""                  ""                  "TanguyBarthelemy" 
#>   [4] ""                  "TanguyBarthelemy"  ""                 
#>   [7] ""                  "charphi"           ""                 
#>  [10] ""                  "TanguyBarthelemy"  "TanguyBarthelemy" 
#>  [13] "TanguyBarthelemy"  "OlesjaL"           ""                 
#>  [16] ""                  ""                  "TanguyBarthelemy" 
#>  [19] ""                  ""                  ""                 
#>  [22] "charphi"           ""                  "Immurb"           
#>  [25] "TanguyBarthelemy"  "TanguyBarthelemy"  ""                 
#>  [28] ""                  ""                  ""                 
#>  [31] ""                  ""                  "charphi"          
#>  [34] ""                  ""                  ""                 
#>  [37] ""                  "palatej"           "palatej"          
#>  [40] ""                  "AQLT"              ""                 
#>  [43] ""                  ""                  "AQLT"             
#>  [46] ""                  "TanguyBarthelemy"  ""                 
#>  [49] "AlessandroPiovani" ""                  "TanguyBarthelemy" 
#>  [52] ""                  "palatej"           "palatej"          
#>  [55] ""                  ""                  ""                 
#>  [58] "MichelleHedger"    "annasmyk"          ""                 
#>  [61] ""                  ""                  ""                 
#>  [64] ""                  "TanguyBarthelemy"  "charphi"          
#>  [67] ""                  ""                  ""                 
#>  [70] ""                  ""                  ""                 
#>  [73] ""                  "a-s-russo"         ""                 
#>  [76] "TanguyBarthelemy"  "Ciamman"           ""                 
#>  [79] ""                  "a-s-russo"         "a-s-russo"        
#>  [82] "a-s-russo"         "CyLuGh"            ""                 
#>  [85] "Immurb"            ""                  ""                 
#>  [88] "Immurb"            ""                  "a-s-russo"        
#>  [91] ""                  "charphi"           "Immurb"           
#>  [94] ""                  ""                  ""                 
#>  [97] "a-s-russo"         "palatej"           "AQLT"             
#> [100] ""                  ""                  ""                 
#> [103] "a-s-russo"         "a-s-russo"         "Immurb"           
#> [106] ""                  ""                  "charphi"          
#> [109] "a-s-russo"         "charphi"           ""                 
#> [112] "palatej"           ""                  ""                 
#> [115] ""                  "a-s-russo"         "m20l22"           
#> [118] "charphi"           "charphi"           ""                 
#> [121] "charphi"           ""                  "charphi"          
#> [124] "AQLT"              ""                  "charphi"          
author_last_comment(all_issues[1L, ])
#> [1] ""