Retrieve information from the issues of GitHub

use gh to ask the API of GitHub and et a list of issues with their labels and milestones.

get_issues(selector, verbose = TRUE, ...)

get_labels(selector, verbose = TRUE, ...)

get_milestones(selector, verbose = TRUE, ...)

Arguments

selector

A SelectorTB object (to define the range of the repo / source used). Created with init_selector().

verbose

A boolean indicating whether to print additional information. Default is TRUE.

...

Other parameters for GitLab as gitlab_url and private_token.

Value

The function get_issues returns an object of class IssuesTB. It is a list composed by object of class IssueTB. An object of class IssueTB represents an issue with simpler structure (with number, title, body and labels).

The function get_labels returns a list representing labels with simpler structure (with name, description, colour).

The function get_milestones returns a list representing milestones with simpler structure (with title, description and due_on).

Details

The functions of get type are useful to retrieve object related to issues from GitHub. So it's possible to retrieve issues, labels and milestones.

The defaults value for the argument dataset_name depends on the function:

defaults is "list_issues.yaml" for get_issues()

defaults is "list_milestones.yaml" for get_milestones()

defaults is "list_labels.yaml" for get_labels()

Examples

# \donttest{
# From online

online_selector <- init_selector(
    source = "GitHub",
    owner = "TanguyBarthelemy",
    repo = "IssueTrackeR"
)
issues <- get_issues(selector = online_selector)
#> Repo: IssueTrackeR  owner: TanguyBarthelemy 
print(issues)
#> There are 14 issues. 
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#53 ──
#> 
#> Title: Coverage v2.1.0
#> Text:
#> To continue the coverage, we put a goal of **90%** for v2.1.0.
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#52 ──
#> 
#> Title: Mutators test coverage for v2.1.0
#> Text:
#> I am setting a coverage target of **30%** for version 2.1.0, using mutators.
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#50 ──
#> 
#> Title: no default for repo and owner and dir
#> Text:
#> BY default, we should set dir to `NULL` or to `""`. So when getting the issues (`get_issues`) we can switch source to "online" if dir is not defined.
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#49 ──
#> 
#> Title: Extract labels and milestones from issues
#> Text:
#> The functions `get_labels` and `get_milestones` should work with `IssueTB` and `IssuesTB` objects.
#> 
#> The goal is to extract all the labels and milestones used in a list of issues or in an issue.
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#47 ──
#> 
#> Title: Bonnes pratiques de documentation
#> Text:
#> Suivre les bonnes pratiques de documentation :
#> 
#> - [ ] Boolean et non Logical
#> - [ ] Suivre le guide écrit pour le rjdverse
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#44 ──
#> 
#> Title: Complete `goodpractice::goodpractice()` checkup
#> Text:
#> Here is the output from {goodpractice}:
#> 
#> ```
#> > gp
#> ── It is good practice to ───────────────────────────────────────────────────────────────────────
#> 
#> ✖ keep functions short and focused. Long functions are harder to understand, test, and maintain.
#>   Consider splitting into smaller helpers. (Default limit of 50 set with o
#> ...
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#43 ──
#> 
#> Title: What to do with type?
#> Text:
#> Une question que je me pose : Si  j'ouvre une issue (mais task, bug ou autre) et que je la ferme avec différent attribut (won't do, duplicated...) qu'est ce qu'il va s'afficher dans l'appel gh::gh ?
#> Aussi est ce que je peux récupérer cette nouvelle info (type) et l'ajouter ? -> que se passe t-il pour les anciennes issu
#> ...
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#36 ──
#> 
#> Title: `get_issue()` missing warning
#> Text:
#> Lorsqu'on appelle `get_issues()` avec repo et owner (et sans source), de base il va dans la local et se fiche de ces arguments:
#> 
#> ```r
#> get_issues(owner = "r-lib", repo = "devtools") |> filter_issues(fields = c("title", "body"), values = "java") 
#> > The issues will be read from C:\Users\UTZK0M\Documents\Projets R\TODO\dat
#> ...
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#31 ──
#> 
#> Title: Coverage v2.0.0
#> Text:
#> To continue the coverage, we put a goal of **80%** for v2.0.0
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#24 ──
#> 
#> Title: Continue coverage
#> Text:
#> The current coverage is 26%. We should update it a lot!
#> 
#> Some functions are not tested:
#> 
#> - [x] `[.IssuesTB`
#> - [ ] `[[<-.IssuesTB`
#> - [ ] `[<-.IssuesTB`
#> - [ ] `add_n_years`
#> - [ ] `append`
#> - [ ] `append.default`
#> - [ ] `append.IssuesTB`
#> - [ ] `author_last_comment`
#> - [ ] `author_last_comment.IssuesTB`
#> - [ ] `author_last_com
#> ...
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#22 ──
#> 
#> Title: Change yaml to parquet
#> Text:
#> Is it possible to add this *new feature*?
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#20 ──
#> 
#> Title: [New feature] Who contributes?
#> Text:
#> Créer un nouveau graphiques / nouvelle statistiques sur la contribution des personnes (ouvrir des issues, répondre, les fermer, mettre des commentaires...)
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#19 ──
#> 
#> Title: Renommer `write_to_datasets` en `write`
#> Text:
#> 
#> 
#> ── Issue TanguyBarthelemy/IssueTrackeR#16 ──
#> 
#> Title: [Bug] `update_database()` needs a new argument for `list_issues.yaml`
#> Text:
#> Hi there, thank you so much for the fantastic package. 
#> 
#> When I download the issues as `get_issues(source = "online", state = "all")` and then write the database to local, I get a `list_issues.yaml` output. However, the `update_database()` function does not have a `list_issues` argument to read this issue table that co
#> ...
#> 

labels <- get_labels(selector = online_selector)
#> Repo: IssueTrackeR  owner: TanguyBarthelemy 
#> Reading labels... Done!
#> 41 labels found.
print(labels)
#> There is 1 repo.
#> - TanguyBarthelemy/IssueTrackeR:, automated issue, communication, complexity: complex, complexity: moderate, complexity: trivial/easy, complexity: very-complex, duplicate, effort: long, effort: medium, effort: short, effort: unknown, effort: very-long, good first issue, good practice, help wanted, invalid, missing documentation, missing examples, missing tests, priority: critical, priority: high, priority: low, priority: medium, question, report, status: can't-reproduce, status: completed, status: duplicate, status: in-progress, status: needs-information, status: on-hold, status: review-needed, status: triage, status: won't-do/fix, type: bug, type: enhancement, type: feature, type: maintenance, type: question, typo, wontfix

milestones <- get_milestones(selector = online_selector)
#> Repo: IssueTrackeR  owner: TanguyBarthelemy 
#> Reading milestones... 
#> 	-  v2.1.0 ... Done!
#> 	-  v2.0.0 ... Done!
#> Done! 2 milestones found.
print(milestones)
#>   source  title description     due_on closed_at          creator state
#> 1 GitHub v2.1.0                   <NA>      <NA> TanguyBarthelemy  open
#> 2 GitHub v2.0.0             2026-09-30      <NA> TanguyBarthelemy  open
#>   nb_issues_open nb_issues_closed         repo            owner
#> 1              7                0 IssueTrackeR TanguyBarthelemy
#> 2              6                5 IssueTrackeR TanguyBarthelemy
#>                                                            url
#> 1 https://github.com/TanguyBarthelemy/IssueTrackeR/milestone/7
#> 2 https://github.com/TanguyBarthelemy/IssueTrackeR/milestone/2
# }

# From local

local_issues_selector <- init_selector(
    source = "Local",
    file = file.path(
        system.file("data_issues", package = "IssueTrackeR"),
        "list_issues.yaml"
    )
)
local_labels_selector <- init_selector(
    source = "Local",
    file = file.path(
        system.file("data_issues", package = "IssueTrackeR"),
        "list_labels.yaml"
    )
)
local_milestones_selector <- init_selector(
    source = "Local",
    file = file.path(
        system.file("data_issues", package = "IssueTrackeR"),
        "list_milestones.yaml"
    )
)

issues <- get_issues(selector = local_issues_selector)
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_issues.yaml.
labels <- get_labels(selector = local_labels_selector)
#> The labels will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_labels.yaml.
milestones <- get_milestones(selector = local_milestones_selector)
#> The milestones will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_milestones.yaml.