Display IssueTB and IssuesTB object

Display IssueTB and IssuesTB with formatted output in the console

# S3 method for class 'IssueTB'
print(x, ...)

# S3 method for class 'IssuesTB'
print(x, ...)

# S3 method for class 'summary.IssueTB'
print(x, ...)

# S3 method for class 'summary.IssuesTB'
print(x, ...)

# S3 method for class 'LabelsTB'
print(x, ...)

# S3 method for class 'summary.LabelsTB'
print(x, ...)

Arguments

x

a IssueTB or IssuesTB object.

...

Unused argument

Value

invisibly (with invisible()) NULL.

Details

This function displays an issue (IssueTB object) or a list of issues (IssuesTB object) with a formatted output.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.

# Display one issue
print(all_issues[1, ])
#> 
#> ── Issue jdemetra/jdplus-main#763 ──
#> 
#> Title: Error loading V2 workspace with fixed week calendar day
#> Text:
#> If I create a workspace in V2 with a National Calendar with a Fixed Week it cannot be imported with the V3. Error loading the calendar with the following error:
#> ```
#> java.lang.NullPointerException: Cannot invoke "String.toUpperCase(java.util.Locale)" because "day" is null
#> 	at internal.toolkit.base.workspace.file.xml.uti
#> ...
#> 

# Display several issues
print(all_issues[1:10, ])
#> There are 10 issues. 
#> 
#> ── Issue jdemetra/jdplus-main#763 ──
#> 
#> Title: Error loading V2 workspace with fixed week calendar day
#> Text:
#> If I create a workspace in V2 with a National Calendar with a Fixed Week it cannot be imported with the V3. Error loading the calendar with the following error:
#> ```
#> java.lang.NullPointerException: Cannot invoke "String.toUpperCase(java.util.Locale)" because "day" is null
#> 	at internal.toolkit.base.workspace.file.xml.uti
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#761 ──
#> 
#> Title: Reloading an xlsx files in Providers create new SA-Item
#> Text:
#> We observe what appears to be a bug.
#> 
#> ### Description
#> 
#> When reloading a data file in the Providers tab, new series are added to the open SA-Processing.
#> 
#> ### Procedure
#> 
#> 1) Open JDemetra+ v3.5.1
#> 2) Create a new WS
#> 3) Create a new SAP
#> 4) Add data from Providers to SAP
#> 5) Add new points to data file
#> 6) Re-add the data file
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#758 ──
#> 
#> Title: Unify period display in GUI (2)
#> Text:
#> Issue #489 is unresolved. 
#> The problem remains in version 3.5.1. 
#> 
#> Also, the SI ratio display is still different and appears with Q1, Q2, Q3, Q4 :
#> 
#> <img width="605" height="255" alt="Image" src="https://github.com/user-attachments/assets/49948a60-a027-4bed-8947-9e9eedc49537" />
#> 
#> ── Issue jdemetra/jdplus-main#752 ──
#> 
#> Title: Add ways to retrieve WS file names and paths
#> Text:
#> It would be very interesting to add ways to retrieve WS file names and paths.
#> 
#> Is it possible to add methods like `getSource` and `setSource` to be able to retrieve the file containing a SA-Processing, a SA-Item or even a Workspace directly from the Java object?
#> 
#> Thank you
#> 
#> ── Issue jdemetra/jdplus-main#751 ──
#> 
#> Title: Structure of a WS
#> Text:
#> Hello, 
#> 
#> I'm opening an issue but it's more of **a discussion**.
#> 
#> First of all, the WS structure contains:
#> 
#> - data (calendar, regressors and data), 
#> - statistical models (specifications) 
#> - and outputs. 
#> 
#> In the context of moving to a cloud-native solution and changing JD+ to a web API, would it make sense to review th
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#734 ──
#> 
#> Title: Can't refresh a SA-Item when WS newly created
#> Text:
#> ### Introduction
#> 
#> In JDemetra+ v3, it's not possible to refresh a SA-Item with **Right click** > **Refresh** when the newly created WS is not saved.
#> It was possible to refresh a new WS in version 2.
#> 
#> But there is A way to refresh a SA-Item with concurrent policy with the GUI in version 3.5.0 with a newly created WS is 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#733 ──
#> 
#> Title: Changing the `domainSpec` doesn't refresh the model
#> Text:
#> ### Intro
#> 
#> In the latest v3.5.1, there is a *breaking change* with v3.5.0 regarding the specification handling.
#> 
#> ### Step to reproduce
#> 
#> 1) Open JDemetra+ v3.5.1
#> 2) Create a new WS
#> 3) Add a series (with for example RSA3 specification)
#> 4) Change the `domainSpec` (**Right click** > **Edit** > **Reference specification** >
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#711 ──
#> 
#> Title: GUI unresponsive when an incorrectly specified file is opened in the Providers p
#> Text:
#> Consider importing a CSV file in the Providers panel and specifying an incorrect date format (such as using 'mm' rather than 'MM', where the former specifies minutes and the latter specifies the month):
#> 
#> <img width="462" height="501" alt="Image" src="https://github.com/user-attachments/assets/8fac7b3d-29cd-4078-81a3-18
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#698 ──
#> 
#> Title: Series naming restrictions
#> Text:
#> This is just a question:
#> 
#> Are there any character restrictions imposed by the JDemetra+ GUI and/or the Cruncher when naming series (that is, when providing a name or ID for series)? For example, are certain special characters not allowed, or are there any character-length limits, and do these differ for the GUI vs the 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#695 ──
#> 
#> Title: Infinity result for diagnostics.fcast-outsample-variance causes remainder of dem
#> Text:
#> I've found a bug with the cruncher. 
#> 
#> Here's the output I saw at the end of my run: 
#> - At the end of the output printed out by cruncher was this: (note the 'failed' line)
#> ```
#> Flushing bundle...
#> Generate Output
#> Csv matrix
#> failed: Input length = 1
#> csv
#> generated
#> Saving new processing...
#> ```
#> - output series were produced a
#> ...
#> 

# Display the summary of one issue
summary(all_issues[2, ])
#> ── Issue jdemetra/jdplus-main#761 ──
#> 
#> Labels: bug
#> State: 🟢 Open
#> Nb comments: 1
#> 
#> Title: Reloading an xlsx files in Providers create new SA-Item
#> Text:
#> We observe what appears to be a bug.
#> 
#> ### Description
#> 
#> When reloading a data file in the Providers tab, new series are added to the open SA-Processing.
#> 
#> ### Procedure
#> 
#> 1) Open JDemetra+ v3.5.1
#> 2) Create a new WS
#> 3) Create a new SAP
#> 4) Add data from Providers to SAP
#> 5) Add new points to data file
#> 6) Re-add the data file under Providers (even if already present and with star)
#> 7) Click on "Reload"
#> 8) Series are added again to SAP
#> 
#> Here's a short video describing the procedure:
#> 
#> https://github.com/user-attachments/assets/323a11ee-9949-467d-8f43-79b4cdf8ab3b
#> 
#> Comments:
#>  
#> Comment n°1 by Immurb:
#> 
#> Just to add some additional information:
#> 
#> 1. Still happens in the develop branch
#> 2. Seems to duplicate only the last added items (last transfer) when the corresponding data file is reloaded
#> 3. Only happens once, so if you reload the same data file multiple times only one set of duplicates is added
#> 4. No new data is needed in the data file and it doesn't matter if you add any data files to the provider (Step 5 and 6)
#> 5. Seems to be related to the events fired in InternalTsProvider (so not only xlsx shows this behaviour)
#>  

# Display the summary of
summary(all_issues[1:10, ])
#> There are 10 issues. 
#> - jdemetra/jdplus-main#763 🟢 Open 
#> - jdemetra/jdplus-main#761 🟢 Open 
#> - jdemetra/jdplus-main#758 🟢 Open 
#> - jdemetra/jdplus-main#752 🟢 Open 
#> - jdemetra/jdplus-main#751 🟢 Open 
#> - jdemetra/jdplus-main#734 🟢 Open 
#> - jdemetra/jdplus-main#733 🟢 Open 
#> - jdemetra/jdplus-main#711 🟢 Open 
#> - jdemetra/jdplus-main#698 🟢 Open 
#> - jdemetra/jdplus-main#695 🟢 Open