Display IssueTB and IssuesTB object

Display IssueTB and IssuesTB with formatted output in the console

# S3 method for class 'IssueTB'
print(x, ...)

# S3 method for class 'IssuesTB'
print(x, ...)

# S3 method for class 'summary.IssueTB'
print(x, ...)

# S3 method for class 'summary.IssuesTB'
print(x, ...)

# S3 method for class 'LabelsTB'
print(x, ...)

# S3 method for class 'summary.LabelsTB'
print(x, ...)

Arguments

x

a IssueTB or IssuesTB object.

...

Unused argument

Value

invisibly (with invisible()) NULL.

Details

This function displays an issue (IssueTB object) or a list of issues (IssuesTB object) with a formatted output.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "list_issues.yaml"
)
#> Looking into list_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_issues.yaml.

# Display one issue
print(all_issues[1, ])
#> 
#> ── Issue jdemetra/jdplus-main#104 ──
#> 
#> Title: Bump org.projectlombok:lombok from 1.18.28 to 1.18.30
#> Text:
#> Bumps [org.projectlombok:lombok](https://github.com/projectlombok/lombok) from 1.18.28 to 1.18.30.
#> <details>
#> <summary>Changelog</summary>
#> <p><em>Sourced from <a href="https://github.com/projectlombok/lombok/blob/master/doc/changelog.markdown">org.projectlombok:lombok's changelog</a>.</em></p>
#> <blockquote>
#> <h3>v1.18.30 
#> ...
#> 

# Display several issues
print(all_issues[1:10, ])
#> There are 10 issues. 
#> 
#> ── Issue jdemetra/jdplus-main#104 ──
#> 
#> Title: Bump org.projectlombok:lombok from 1.18.28 to 1.18.30
#> Text:
#> Bumps [org.projectlombok:lombok](https://github.com/projectlombok/lombok) from 1.18.28 to 1.18.30.
#> <details>
#> <summary>Changelog</summary>
#> <p><em>Sourced from <a href="https://github.com/projectlombok/lombok/blob/master/doc/changelog.markdown">org.projectlombok:lombok's changelog</a>.</em></p>
#> <blockquote>
#> <h3>v1.18.30 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#236 ──
#> 
#> Title: Fix drag&drop from Excel
#> Text:
#> 
#> 
#> ── Issue jdemetra/jdplus-main#238 ──
#> 
#> Title: Missing chart export to SVG
#> Text:
#> JD+ v2 is able to export a chart to SVG but v3 is not.
#> 
#> ── Issue jdemetra/jdplus-main#51 ──
#> 
#> Title: theil2 slightly adapted
#> Text:
#> 
#> 
#> ── Issue jdemetra/jdplus-main#590 ──
#> 
#> Title: Summary info on pre-specified outliers not clear
#> Text:
#> In the JD+ GUI, the user can insert 'pre-specified outliers' as shown below, and these can be fixed or non-fixed:
#> 
#> ![Image](https://github.com/user-attachments/assets/d4a65a9a-3173-4e67-80a3-1e1f4cdd7ee7)
#> 
#> If the user then looks at the **Summary** section of the **Main **results**** section in the output area for the s
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#38 ──
#> 
#> Title: Improve CsvInformationFormatter
#> Text:
#> 
#> 
#> ── Issue jdemetra/jdplus-main#610 ──
#> 
#> Title: Mode is 'Multiplicative' when transform function is 'None'
#> Text:
#> My additive series has regarima.transform.function = 'None', but x11.mode is 'Multiplicative':
#> ![Image](https://github.com/user-attachments/assets/6934cd41-6613-43d7-8903-3f2deb923f70)
#> 
#> This is the same series with regarima.transform.function = 'Log', and x11.mode is also 'Multiplicative':
#> ![Image](https://github.com/u
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#658 ──
#> 
#> Title: Migrate protobuf-maven-plugin from org.xolstice.maven.plugins to io.github.ascop
#> Text:
#> From https://github.com/xolstice/protobuf-maven-plugin/ 
#> 
#> > This project is no longer being maintained, and no further contributions will be considered.
#> > 
#> > Please consider switching to alternatives, such as:
#> > https://github.com/ascopes/protobuf-maven-plugin
#> > 
#> 
#> ── Issue jdemetra/jdplus-main#57 ──
#> 
#> Title: Bump protobuf-java from 3.21.12 to 3.23.2
#> Text:
#> Bumps [protobuf-java](https://github.com/protocolbuffers/protobuf) from 3.21.12 to 3.23.2.
#> <details>
#> <summary>Commits</summary>
#> <ul>
#> <li><a href="https://github.com/protocolbuffers/protobuf/commit/a74f54b724bdc2fe0bfc271f4dc0ceb159805625"><code>a74f54b</code></a> Updating version.json and repo version numbers to: 23.2<
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#112 ──
#> 
#> Title: JWSACruncher and parameters `fullseriesname` and `refreshall`
#> Text:
#> I tried the following parameters files an a workspace with data created by an Excel file and it has no effect. Are the parameters `fullseriesname` and `refreshall` suppose to work in the JWSACruncher? Am I doing something wrong?
#> ```
#> <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
#> <wsaConfig bundle="10000" cs
#> ...
#> 

# Display the summary of one issue
summary(all_issues[2, ])
#> Error in if (object$has_labels) {    object$labels_name <- object$labels$name    object$labels_bgcolor <- object$labels$color    object$labels_color <- c("grey8", "ivory")[isDark(object$labels_bgcolor) +         1L]    object$labels_url <- paste("https://github.com", object$owner,         object$repo, "labels", utils::URLencode(object$labels_name),         sep = "/")}: argument is of length zero

# Display the summary of
summary(all_issues[1:10, ])
#> There are 10 issues. 
#> - jdemetra/jdplus-main#104 🟢 Open 
#> - jdemetra/jdplus-main#236 🟢 Open 
#> - jdemetra/jdplus-main#238 ✔ Completed 
#> - jdemetra/jdplus-main#51 🟢 Open 
#> - jdemetra/jdplus-main#590 🟢 Open 
#> - jdemetra/jdplus-main#38 🟢 Open 
#> - jdemetra/jdplus-main#610 ✔ Completed 
#> - jdemetra/jdplus-main#658 ✔ Completed 
#> - jdemetra/jdplus-main#57 🟢 Open 
#> - jdemetra/jdplus-main#112 ♻ Re-opened