Display IssueTB and IssuesTB object

Display IssueTB and IssuesTB with formatted output in the console

# S3 method for class 'IssueTB'
print(x, ...)

# S3 method for class 'IssuesTB'
print(x, ...)

# S3 method for class 'summary.IssueTB'
print(x, ...)

# S3 method for class 'summary.IssuesTB'
print(x, ...)

# S3 method for class 'LabelsTB'
print(x, ...)

# S3 method for class 'summary.LabelsTB'
print(x, ...)

Arguments

x

a IssueTB or IssuesTB object.

...

Unused argument

Value

invisibly (with invisible()) NULL.

Details

This function displays an issue (IssueTB object) or a list of issues (IssuesTB object) with a formatted output.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.

# Display one issue
print(all_issues[1, ])
#> 
#> ── Issue jdemetra/jdplus-main#1006 ──
#> 
#> Title: Benchmarking Results includes forecast observations beyond the input series
#> Text:
#> <img width="340" height="216" alt="Image" src="https://github.com/user-attachments/assets/cd77b619-d644-4d03-925e-6a858eeb3709" />
#> 
#> While testing JDemetra+ 3.7.1, I noticed an unexpected behaviour in the **Benchmarking Results** output. The input series used in this example ends in **2025 Q1**. However, the table displ
#> ...
#> 

# Display several issues
print(all_issues[1:10, ])
#> There are 10 issues. 
#> 
#> ── Issue jdemetra/jdplus-main#1006 ──
#> 
#> Title: Benchmarking Results includes forecast observations beyond the input series
#> Text:
#> <img width="340" height="216" alt="Image" src="https://github.com/user-attachments/assets/cd77b619-d644-4d03-925e-6a858eeb3709" />
#> 
#> While testing JDemetra+ 3.7.1, I noticed an unexpected behaviour in the **Benchmarking Results** output. The input series used in this example ends in **2025 Q1**. However, the table displ
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#1004 ──
#> 
#> Title: Outlier type and position are missing in Matrix |> Outliers
#> Text:
#> Related to cosa-tsa-shop/UserSupport#10
#> 
#> In the section Matrix |> Outliers (from a SA-Processing), the version 2 and 3 differs by their presentation:
#> 
#> - In version 2 we observe a column by outliers with type, position, coeff and T-Stats
#> - In version 3 we observe 2 columns by outliers with coeffs and T-Stats
#> 
#> Is it poss
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#987 ──
#> 
#> Title: The variable `regression.td-ftest` is missing in JD+ output
#> Text:
#> I generate a `demetra_m` output using the graphical interface (version 3.7.1) and I notice that the `td-ftest` column (`regression.td-ftest` item) is completely empty.
#> Is this a connection error between the input and output?
#> However, in version 2 (I tested version 2.2.6), for the same specifications and the same series
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#962 ──
#> 
#> Title: JDemetra+ doesn't import duplicated variables
#> Text:
#> If two variables have the same value, only the last one will be read.
#> 
#> Here is an example in R:
#> 
#> ``` r
#> library("rjd3workspace")
#> path_xml <- tempfile(fileext = ".xml")
#> writeLines(
#>     text = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
#> <tsVariables xmlns="ec/tss.core">
#>     <tsVariable name="VAR1">
#>         <t
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#961 ──
#> 
#> Title: Calendario - JDemetra V3
#> Text:
#> Buenos días. 
#> En la versión 3 de JJDemetra+ intento cargar el calendario para Argentina creado por mi, pero no lo reconoce la aplicación. El calendario se encuentra en formato cfgx y podía agregarlo en las versiones 2.0. De no agregar en las versiones 3, ¿cuál sería la forma para agregar los calendarios particulares po
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#943 ──
#> 
#> Title: Generating output only if the SAP is computed
#> Text:
#> I try to understand why the output generation wors with the GUI sometimes and sometimes not.
#> 
#> In this issue, I will use the verbs "compute" and "estimate" to refer to the same thing: obtaining the results of the SA estimate for my series. But perhaps there is a fundamental difference between the two actions...
#> 
#> ## Exam
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#940 ──
#> 
#> Title: How the GUI sort by specification?
#> Text:
#> In the GUI, it's possible to sort the SA-Item of a SA-Processing by **Reference spec**. But how is it handle?
#> 
#> I notice (in this example) that this isn’t done by name and that the `^` and `v` keys aren’t used to reverse the direction.
#> 
#> https://github.com/user-attachments/assets/c627e166-e28b-4de2-818f-821dd96c6b6c
#> 
#> ── Issue jdemetra/jdplus-main#912 ──
#> 
#> Title: Add JDemetra+ to the EU Open Source Solutions Catalogue
#> Text:
#> > The EU Open Source Software Solutions Catalogue (EU OSS Catalogue) has been created by the FOSSEPS Preparatory initiative, [Free and Open Source Software for European Public Services](https://interoperable-europe.ec.europa.eu/collection/fosseps), which aims to promote cooperation at European Union level by giving vis
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#888 ──
#> 
#> Title: Is it possible to choose the value `(?)` in GUI output?
#> Text:
#> The aim of this issue is to find out whether, when I generate an excel output with the GUI, it's possible to choose the length of series which items ends with `(?)`.
#> 
#> I will take the example of the forecasted series. 
#> 
#> In X13-RegARIMA, there are forecasts in the preprocessing part but only with `(?)` items and it's not
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#883 ──
#> 
#> Title: Adding a button to "refresh" a WS
#> Text:
#> *related to #882*
#> 
#> I'm **not** talking here about the refresh policy that can be applied to a SA-Item or to a SA-Processing. 
#> I'm talking about the action that is executed each time a workspace is saved or when it is closed and reopened. The WS is saved in a **XML format** but some information are also updated (maybe i
#> ...
#> 

# Display the summary of one issue
summary(all_issues[2, ])
#> ── Issue jdemetra/jdplus-main#1004 ──
#> 
#> Labels: 
#> State: 🟢 Open
#> Nb comments: 0
#> 
#> Title: Outlier type and position are missing in Matrix |> Outliers
#> Text:
#> Related to cosa-tsa-shop/UserSupport#10
#> 
#> In the section Matrix |> Outliers (from a SA-Processing), the version 2 and 3 differs by their presentation:
#> 
#> - In version 2 we observe a column by outliers with type, position, coeff and T-Stats
#> - In version 3 we observe 2 columns by outliers with coeffs and T-Stats
#> 
#> Is it possible to include the outliers type and position in version 3?
#> 
#> Here is the comparison:
#> 
#> <img width="1920" height="1031" alt="Image" src="https://github.com/user-attachments/assets/825ba55f-8f7a-4309-ba9d-0ead44f26889" />
#> 
#> 

# Display the summary of
summary(all_issues[1:10, ])
#> There are 10 issues. 
#> - jdemetra/jdplus-main#1006 🟢 Open 
#> - jdemetra/jdplus-main#1004 🟢 Open 
#> - jdemetra/jdplus-main#987 🟢 Open 
#> - jdemetra/jdplus-main#962 🟢 Open 
#> - jdemetra/jdplus-main#961 🟢 Open 
#> - jdemetra/jdplus-main#943 🟢 Open 
#> - jdemetra/jdplus-main#940 🟢 Open 
#> - jdemetra/jdplus-main#912 ♻ Re-opened 
#> - jdemetra/jdplus-main#888 🟢 Open 
#> - jdemetra/jdplus-main#883 🟢 Open