Compute a summary of an issue or a list of issues

Compute a summary of an issue or a list of issues

# S3 method for class 'IssueTB'
summary(object, ...)

# S3 method for class 'IssuesTB'
summary(object, ...)

Arguments

object

a IssueTB or IssuesTB object.

...

Unused argument

Value

invisibly (with invisible()) NULL.

Details

This function compute the summary of an issue (IssueTB object) with adding some information (number of comments, ...). For a list of issues (IssuesTB object), it just summarise the information with statistics by modalities.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "list_issues.yaml"
)
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/list_issues.yaml.

# Summarise one issue
summary(all_issues[1, ])
#> 
#> ── Issue jdemetra/jdplus-main#104 ──
#> 
#> Labels: dependencies, java
#> State: 🟢 Open
#> Nb comments: 1
#> 
#> Title: Bump org.projectlombok:lombok from 1.18.28 to 1.18.30
#> Text:
#> Bumps [org.projectlombok:lombok](https://github.com/projectlombok/lombok) from 1.18.28 to 1.18.30.
#> <details>
#> <summary>Changelog</summary>
#> <p><em>Sourced from <a href="https://github.com/projectlombok/lombok/blob/master/doc/changelog.markdown">org.projectlombok:lombok's changelog</a>.</em></p>
#> <blockquote>
#> <h3>v1.18.30 (September 20th, 2023)</h3>
#> <ul>
#> <li>PLATFORM: Initial JDK21 support added. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3393">#3393</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3393">projectlombok/lombok#3393</a>).</li>
#> <li>BUGFIX: Any <code>@Helper</code> class directly in a method (and not nested more deeply) wouldn't work. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3370">#3370</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3370">projectlombok/lombok#3370</a>).</li>
#> <li>BUGFIX: If using the module system and lombok is on the runtime classpath (shouldn't be, but happens), you'd get a split package error: <code>Package org.objectweb.asm in both module lombok and module org.objectweb.asm</code>. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3474">#3474</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3474">projectlombok/lombok#3474</a>)</li>
#> <li>BUGFIX: Lombok wasn't properly copying the annotations it should be copying when generating methods in <code>record</code>s. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3315">#3315</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3315">projectlombok/lombok#3315</a>).</li>
#> <li>BUGFIX: Delomboking anything with <code>@lombok.Singular</code> in it wouldn't remove that annotation. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/1377">#1377</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/1377">projectlombok/lombok#1377</a>).</li>
#> <li>BUGFIX: Calling extension methods such that automatic widening is applied (i.e. calling <code>void ext(long arg)</code> with an <code>int</code>) would fail at runtime. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3463">#3463</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3463">projectlombok/lombok#3463</a>).</li>
#> <li>BUGFIX: Extension methods can now be used in records. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3450">#3450</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3450">projectlombok/lombok#3450</a>).</li>
#> <li>BUGFIX: <code>@Getter(lazy=true)</code> with complicated initialization expressions would fail on javac. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3314">#3314</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3314">projectlombok/lombok#3314</a>).</li>
#> <li>BUGFIX: Using the maven surefire plugin with a <code>module-info.java</code> based project would fail with a <code>SurefireBooterForkException</code>. [Issue <a href="https://redirect.github.com/projectlombok/lombok/issues/3474">#3474</a>](<a href="https://redirect.github.com/projectlombok/lombok/issues/3474">projectlombok/lombok#3474</a>).</li>
#> </ul>
#> </blockquote>
#> </details>
#> <details>
#> <summary>Commits</summary>
#> <ul>
#> <li><a href="https://github.com/projectlombok/lombok/commit/b118bea124bc5aef59cb0c5cac4a086e1e3d98f2"><code>b118bea</code></a> [release] pre-release version bump</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/7cbb5e9d4ad2a8e308711d0a9c59fc4f2c5292a8"><code>7cbb5e9</code></a> [fixes <a href="https://redirect.github.com/projectlombok/lombok/issues/3182">#3182</a>] [docs] Add a note about wonky equality definitions in the core ...</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/9bcaaafed4646ae750bf8d6ca7e3b1c549dc41f1"><code>9bcaaaf</code></a> [fixes <a href="https://redirect.github.com/projectlombok/lombok/issues/3474">#3474</a>] lombok.jar now contains zero paths that could clash with other ...</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/9ef9b58334fe5f861b44a0f76875f54fd24af20b"><code>9ef9b58</code></a> Add JDK 21 to test workflow</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/9f350ac78c50983141508339eb44f9989317be43"><code>9f350ac</code></a> Merge pull request <a href="https://redirect.github.com/projectlombok/lombok/issues/3503">#3503</a> from Rawi01/javac6-singular-test</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/e226cd357837e36c48a117793de3f91bbc644379"><code>e226cd3</code></a> Exclude failing test for javac &lt; 8</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/67cd7a6218d0989656674689e1795bb245470972"><code>67cd7a6</code></a> Merge pull request <a href="https://redirect.github.com/projectlombok/lombok/issues/3494">#3494</a> from Rawi01/enum-getter</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/6623e12fb221e5773beb0175f1a692dbb6e00074"><code>6623e12</code></a> Merge pull request <a href="https://redirect.github.com/projectlombok/lombok/issues/3486">#3486</a> from Rawi01/superbuilder-javadoc</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/28e5ddb075b61525ba686f3ed6ecbf7f762fc558"><code>28e5ddb</code></a> <a href="https://redirect.github.com/projectlombok/lombok/issues/3314">#3314</a> Add line to changelog</li>
#> <li><a href="https://github.com/projectlombok/lombok/commit/60ed4adfcee2416d757d312aa63281959ef06a9b"><code>60ed4ad</code></a> Merge pull request <a href="https://redirect.github.com/projectlombok/lombok/issues/3482">#3482</a> from Rawi01/getter-lazy</li>
#> <li>Additional commits viewable in <a href="https://github.com/projectlombok/lombok/compare/v1.18.28...v1.18.30">compare view</a></li>
#> </ul>
#> </details>
#> <br />
#> 
#> 
#> [![Dependabot compatibility score](https://dependabot-badges.githubapp.com/badges/compatibility_score?dependency-name=org.projectlombok:lombok&package-manager=maven&previous-version=1.18.28&new-version=1.18.30)](https://docs.github.com/en/github/managing-security-vulnerabilities/about-dependabot-security-updates#about-compatibility-scores)
#> 
#> Dependabot will resolve any conflicts with this PR as long as you don't alter it yourself. You can also trigger a rebase manually by commenting `@dependabot rebase`.
#> 
#> [//]: # (dependabot-automerge-start)
#> [//]: # (dependabot-automerge-end)
#> 
#> ---
#> 
#> <details>
#> <summary>Dependabot commands and options</summary>
#> <br />
#> 
#> You can trigger Dependabot actions by commenting on this PR:
#> - `@dependabot rebase` will rebase this PR
#> - `@dependabot recreate` will recreate this PR, overwriting any edits that have been made to it
#> - `@dependabot merge` will merge this PR after your CI passes on it
#> - `@dependabot squash and merge` will squash and merge this PR after your CI passes on it
#> - `@dependabot cancel merge` will cancel a previously requested merge and block automerging
#> - `@dependabot reopen` will reopen this PR if it is closed
#> - `@dependabot close` will close this PR and stop Dependabot recreating it. You can achieve the same result by closing it manually
#> - `@dependabot show <dependency name> ignore conditions` will show all of the ignore conditions of the specified dependency
#> - `@dependabot ignore this major version` will close this PR and stop Dependabot creating any more for this major version (unless you reopen the PR or upgrade to it yourself)
#> - `@dependabot ignore this minor version` will close this PR and stop Dependabot creating any more for this minor version (unless you reopen the PR or upgrade to it yourself)
#> - `@dependabot ignore this dependency` will close this PR and stop Dependabot creating any more for this dependency (unless you reopen the PR or upgrade to it yourself)
#> 
#> 
#> </details>
#> 
#> Comments:
#>  
#> Comment n°1 by charphi:
#> 
#> @dependabot rebase
#>  

# Summarise several issues
summary(all_issues[1:10, ])
#> There are 10 issues: 
#> - jdemetra/jdplus-main#104 🟢 Open 
#> - jdemetra/jdplus-main#236 🟢 Open 
#> - jdemetra/jdplus-main#238 ✔ Completed 
#> - jdemetra/jdplus-main#51 🟢 Open 
#> - jdemetra/jdplus-main#590 🟢 Open 
#> - jdemetra/jdplus-main#38 🟢 Open 
#> - jdemetra/jdplus-main#610 ✔ Completed 
#> - jdemetra/jdplus-main#658 ✔ Completed 
#> - jdemetra/jdplus-main#57 🟢 Open 
#> - jdemetra/jdplus-main#112 ♻ Re-opened