Compute a summary of an issue or a list of issues

Compute a summary of an issue or a list of issues

# S3 method for class 'IssueTB'
summary(object, ...)

# S3 method for class 'IssuesTB'
summary(object, with_labels = FALSE, ...)

# S3 method for class 'LabelsTB'
summary(object, ...)

Arguments

object

a IssueTB or IssuesTB object.

...

Unused argument

with_labels

Boolean. Display the labels with the list of issues. Default is FALSE.

Value

invisibly (with invisible()) NULL.

Details

This function compute the summary of an issue (IssueTB object) with adding some information (number of comments, ...). For a list of issues (IssuesTB object), it just summarise the information with statistics by modalities.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.

# Summarise one issue
summary(all_issues[1, ])
#> 
#> ── Issue jdemetra/jdplus-main#1006 ──
#> 
#> Labels: 
#> State: 🟢 Open
#> Nb comments: 0
#> 
#> Title: Benchmarking Results includes forecast observations beyond the input series
#> Text:
#> <img width="340" height="216" alt="Image" src="https://github.com/user-attachments/assets/cd77b619-d644-4d03-925e-6a858eeb3709" />
#> 
#> While testing JDemetra+ 3.7.1, I noticed an unexpected behaviour in the **Benchmarking Results** output. The input series used in this example ends in **2025 Q1**. However, the table displayed under **Benchmarking Results** contains additional observations beyond the end of the available data. These values appear to be forecasted values generated internally by the benchmarking procedure.
#> 
#> Although the forecasts themselves may be valid, they are displayed in the same table as the observed data without any visual distinction. As a result, it is difficult for users to identify where the original series ends and where the forecast period begins.
#> 
#> Version: JDemetra+ 3.7.1
#> 
#> 
#> 
#> 

# Summarise several issues
summary(all_issues[1:10, ])
#> There are 10 issues. 
#> - jdemetra/jdplus-main#1006 🟢 Open 
#> - jdemetra/jdplus-main#1004 🟢 Open 
#> - jdemetra/jdplus-main#987 🟢 Open 
#> - jdemetra/jdplus-main#962 🟢 Open 
#> - jdemetra/jdplus-main#961 🟢 Open 
#> - jdemetra/jdplus-main#943 🟢 Open 
#> - jdemetra/jdplus-main#940 🟢 Open 
#> - jdemetra/jdplus-main#912 ♻ Re-opened 
#> - jdemetra/jdplus-main#888 🟢 Open 
#> - jdemetra/jdplus-main#883 🟢 Open