Update database

Update the different local database (issues, labels and milestones) with the online reference.

update_database(
  dataset_dir = getOption("IssueTrackeR.dataset.dir"),
  datasets_name = c(open = "open_issues.yaml", closed = "closed_issues.yaml", labels =
    "list_labels.yaml", milestones = "list_milestones.yaml"),
  verbose = TRUE,
  ...
)

Arguments

dataset_dir

A character string specifying the path which contains the datasets (only taken into account if source is set to "local"). Defaults to the package option IssueTrackeR.dataset.dir.

datasets_name

A named character string of length 4, specifying the names of the different datasets which will be written. The names datasets_name have to be "open", "closed", "labels" and "milestones". Defaults to c(open = "open_issues.yaml", closed = "closed_issues.yaml", labels = "list_labels.yaml", milestones = "list_milestones.yaml") .

verbose

A logical value indicating whether to print additional information. Default is TRUE.

...

Additional arguments for connecting to the GitHub repository:

repo A character string specifying the GitHub repository name. Defaults to the package option IssueTrackeR.repo.

owner A character string specifying the GitHub owner. Defaults to the package option IssueTrackeR.owner. (See the documentation of get to have more information on theses parameters):

Value

invisibly (with invisible()) TRUE.

Examples

# \donttest{
update_database()
#> Reading issues...
#> Issue n° 1... Done!
#> Issue n° 2... Done!
#> Issue n° 3... Done!
#> Issue n° 4... Done!
#> 4 issues found.
#> The datasets will be exported to /tmp/RtmpKJvpw0/data/open_issues.yaml.
#> The file already exists and will be overwritten.
#> Reading issues...
#> Issue n° 1... Done!
#> Issue n° 2... Done!
#> Issue n° 3... Done!
#> Issue n° 4... Done!
#> Issue n° 5... Done!
#> Issue n° 6... Done!
#> Issue n° 7... Done!
#> Issue n° 8... Done!
#> Issue n° 9... Done!
#> Issue n° 10... Done!
#> Issue n° 11... Done!
#> Issue n° 12... Done!
#> Issue n° 13... Done!
#> Issue n° 14... Done!
#> Issue n° 15... Done!
#> Issue n° 16... Done!
#> Issue n° 17... Done!
#> Issue n° 18... Done!
#> Issue n° 19... Done!
#> Issue n° 20... Done!
#> Issue n° 21... Done!
#> Issue n° 22... Done!
#> Issue n° 23... Done!
#> Issue n° 24... Done!
#> Issue n° 25... Done!
#> Issue n° 26... Done!
#> Issue n° 27... Done!
#> Issue n° 28... Done!
#> Issue n° 29... Done!
#> Issue n° 30... Done!
#> Issue n° 31... Done!
#> Issue n° 32... Done!
#> Issue n° 33... Done!
#> Issue n° 34... Done!
#> Issue n° 35... Done!
#> Issue n° 36... Done!
#> Issue n° 37... Done!
#> Issue n° 38... Done!
#> Issue n° 39... Done!
#> Issue n° 40... Done!
#> Issue n° 41... Done!
#> Issue n° 42... Done!
#> Issue n° 43... Done!
#> Issue n° 44... Done!
#> Issue n° 45... Done!
#> Issue n° 46... Done!
#> Issue n° 47... Done!
#> Issue n° 48... Done!
#> Issue n° 49... Done!
#> Issue n° 50... Done!
#> Issue n° 51... Done!
#> Issue n° 52... Done!
#> Issue n° 53... Done!
#> Issue n° 54... Done!
#> Issue n° 55... Done!
#> Issue n° 56... Done!
#> Issue n° 57... Done!
#> Issue n° 58... Done!
#> Issue n° 59... Done!
#> Issue n° 60... Done!
#> Issue n° 61... Done!
#> Issue n° 62... Done!
#> Issue n° 63... Done!
#> Issue n° 64... Done!
#> Issue n° 65... Done!
#> Issue n° 66... Done!
#> Issue n° 67... Done!
#> Issue n° 68... Done!
#> Issue n° 69... Done!
#> Issue n° 70... Done!
#> Issue n° 71... Done!
#> Issue n° 72... Done!
#> Issue n° 73... Done!
#> Issue n° 74... Done!
#> Issue n° 75... Done!
#> Issue n° 76... Done!
#> Issue n° 77... Done!
#> Issue n° 78... Done!
#> Issue n° 79... Done!
#> Issue n° 80... Done!
#> Issue n° 81... Done!
#> Issue n° 82... Done!
#> Issue n° 83... Done!
#> Issue n° 84... Done!
#> Issue n° 85... Done!
#> Issue n° 86... Done!
#> Issue n° 87... Done!
#> Issue n° 88... Done!
#> Issue n° 89... Done!
#> Issue n° 90... Done!
#> Issue n° 91... Done!
#> Issue n° 92... Done!
#> Issue n° 93... Done!
#> Issue n° 94... Done!
#> Issue n° 95... Done!
#> Issue n° 96... Done!
#> Issue n° 97... Done!
#> Issue n° 98... Done!
#> Issue n° 99... Done!
#> Issue n° 100... Done!
#> Issue n° 101... Done!
#> Issue n° 102... Done!
#> Issue n° 103... Done!
#> Issue n° 104... Done!
#> Issue n° 105... Done!
#> Issue n° 106... Done!
#> Issue n° 107... Done!
#> Issue n° 108... Done!
#> Issue n° 109... Done!
#> Issue n° 110... Done!
#> Issue n° 111... Done!
#> Issue n° 112... Done!
#> Issue n° 113... Done!
#> Issue n° 114... Done!
#> Issue n° 115... Done!
#> Issue n° 116... Done!
#> Issue n° 117... Done!
#> Issue n° 118... Done!
#> Issue n° 119... Done!
#> Issue n° 120... Done!
#> Issue n° 121... Done!
#> Issue n° 122... Done!
#> Issue n° 123... Done!
#> Issue n° 124... Done!
#> Issue n° 125... Done!
#> Issue n° 126... Done!
#> Issue n° 127... Done!
#> Issue n° 128... Done!
#> Issue n° 129... Done!
#> Issue n° 130... Done!
#> Issue n° 131... Done!
#> Issue n° 132... Done!
#> Issue n° 133... Done!
#> Issue n° 134... Done!
#> Issue n° 135... Done!
#> Issue n° 136... Done!
#> Issue n° 137... Done!
#> Issue n° 138... Done!
#> Issue n° 139... Done!
#> Issue n° 140... Done!
#> Issue n° 141... Done!
#> Issue n° 142... Done!
#> Issue n° 143... Done!
#> Issue n° 144... Done!
#> Issue n° 145... Done!
#> Issue n° 146... Done!
#> Issue n° 147... Done!
#> Issue n° 148... Done!
#> Issue n° 149... Done!
#> Issue n° 150... Done!
#> Issue n° 151... Done!
#> Issue n° 152... Done!
#> Issue n° 153... Done!
#> Issue n° 154... Done!
#> Issue n° 155... Done!
#> Issue n° 156... Done!
#> Issue n° 157... Done!
#> Issue n° 158... Done!
#> Issue n° 159... Done!
#> 159 issues found.
#> The datasets will be exported to /tmp/RtmpKJvpw0/data/closed_issues.yaml.
#> The file already exists and will be overwritten.
#> Reading labels... Done!
#> 8 labels found.
#> The datasets will be exported to /tmp/RtmpKJvpw0/data/list_labels.yaml.
#> The file already exists and will be overwritten.
#> Reading milestones... 
#> Done! 0 milestones found.
#> The datasets will be exported to /tmp/RtmpKJvpw0/data/list_milestones.yaml.
#> The file already exists and will be overwritten.
# }