Check for comments in GitHub Issues

Function to filter issues with (or without) comments.

with_comments(x, ...)

# S3 method for class 'IssuesTB'
with_comments(x, negate = FALSE, ...)

Arguments

x

An object of class IssuesTB.

...

Currently not used.

negate

boolean indicating if we are searching for issues WITHOUT comments. Default is FALSE.

Value

An object IssuesTB with issues that satisfy the condition.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.
with_comments(all_issues)
#> There are 58 issues. 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#1039 ──
#> 
#> Title: Older WS are read not correctly
#> Text:
#> I created a WS with the version **3.7.1** with 1 SA-Processing with 1 SA-Item with:
#> 
#> - the series AirPassengers ([air.csv](https://github.com/user-attachments/files/30928390/air.csv))
#> - the specification:
#>   - X13 RSA3
#>   - 1 outliers (AO in 1956-08-01) in the estimation spec
#>   - a basic span (series span) from 1949-01-0
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#1023 ──
#> 
#> Title: OpenSSF badge
#> Text:
#> I recently discovered the OpenSSF badge.
#> 
#> I was wondering if it applied to JDemetra+. It’s a badge that recognises packages that follow open-source best practice.
#> 
#> The link is https://www.bestpractices.dev/
#> 
#> ── Issue jdemetra/jdplus-main#987 ──
#> 
#> Title: The variable `regression.td-ftest` is missing in JD+ output
#> Text:
#> I generate a `demetra_m` output using the graphical interface (version 3.7.1) and I notice that the `td-ftest` column (`regression.td-ftest` item) is completely empty.
#> Is this a connection error between the input and output?
#> However, in version 2 (I tested version 2.2.6), for the same specifications and the same series
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#961 ──
#> 
#> Title: Calendario - JDemetra V3
#> Text:
#> Buenos días. 
#> En la versión 3 de JJDemetra+ intento cargar el calendario para Argentina creado por mi, pero no lo reconoce la aplicación. El calendario se encuentra en formato cfgx y podía agregarlo en las versiones 2.0. De no agregar en las versiones 3, ¿cuál sería la forma para agregar los calendarios particulares po
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#912 ──
#> 
#> Title: Add JDemetra+ to the EU Open Source Solutions Catalogue
#> Text:
#> > The EU Open Source Software Solutions Catalogue (EU OSS Catalogue) has been created by the FOSSEPS Preparatory initiative, [Free and Open Source Software for European Public Services](https://interoperable-europe.ec.europa.eu/collection/fosseps), which aims to promote cooperation at European Union level by giving vis
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#881 ──
#> 
#> Title: Reappearance of a deleted SA-Item
#> Text:
#> I thought I already open an issue about this subject but I can't find any trace of it in older issues...
#> 
#> ### Situation
#> 
#> Some SA-Items that have been deleted reappear when I add other series.
#> 
#> ### Steps to reproduce
#> 
#> 1) Create a new WS
#> 2) Create a SA-Processing
#> 3) Add SA-items
#> 4) Remove one of the SA-Items
#> 5) Add a new
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#878 ──
#> 
#> Title: How is computed the Non parametric (Kruskal-Wallis) test?
#> Text:
#> The function `rjd3toolkit::seasonality_kruskalwallis()` compute the Kruskal-Wallis test. But how is it implemented in x13 call?
#> 
#> Here is a small example. I get two different result if I want to compute it directly from the raw data and get it from the user defined variables:
#> 
#> ``` r
#> library("rjd3x13")
#> #> 
#> #> Attaching p
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#872 ──
#> 
#> Title: Missing parameters
#> Text:
#> Hello,
#> 
#> I am using JDemetra+ 3.6.0 with jwsacruncher 3.6.0, and I noticed that the CSV matrix output (demetra_m) in JD3 is missing some parameters and diagnostics that were present in the CSV matrix output in JD2 versions.
#> With JD3, the demetra_m is only partially filled, several parameters and diagnostics visible in t
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#867 ──
#> 
#> Title: Refresh a model without refreshing the data
#> Text:
#> Currently, to call a refresh policy (in the GUI or the cruncher), the ts metadata of a SA-Item have to be filled and accurate.
#> 
#> But is there a way to refresh a model (with any refresh policy) but without refreshing the data (so just applying the domain and estimation spec with the current data in the SA-Processing or p
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#865 ──
#> 
#> Title: `(frozen)` in SAI names
#> Text:
#> I just created a new WS with csv and Excel files. 
#> When I try to read it from R, I get `(frozen)` in the name of the SA-Items:
#> 
#> https://github.com/user-attachments/assets/28fffeb1-b58c-42d5-8219-c440d30248d9
#> 
#> Here is the R code:
#> 
#> ``` r
#> library("rjd3workspace")
#> 
#> jws <- jws_open("~/tmp.xml")
#> jsap <- jws_sap(jws, 1L)
#> sap_
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#863 ──
#> 
#> Title: Migrate to Java 25
#> Text:
#> The current Java version used by JD+ is Java 17.
#> A new Java LTS has been released and brings significant improvements in performance, language and libraries.
#> 
#> JD+ mostly works on Java 25 but there are a few issues.
#> 
#> Related issue: #160 
#> 
#> ── Issue jdemetra/jdplus-main#859 ──
#> 
#> Title: Cannot add a SA-Item with same name
#> Text:
#> I try to add a second SA-Item from the same file.
#> In the first case, 
#> 
#> - I create an SA-Processing
#> - I first add my SA-Item (here the RF1011 series) and a second series
#> - I try to add the series a second time, but it fails (even if I close and re-open the SA-Processing)
#> - I delete the SA-Item RF1011 and **I try to add 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#858 ──
#> 
#> Title: Error when put TC outliers outside of estimation span
#> Text:
#> When I change both the estimation span (model span) and set a TC outlier outside this span, I get an error :
#> 
#> `No information for this item`
#> 
#> In the GUI:
#> 
#> https://github.com/user-attachments/assets/22b15752-c566-4d38-a7cf-a066525a160d
#> 
#> 
#> And here is a small reproducible example with R packages {rjd3toolkit} and {rjd3x13
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#856 ──
#> 
#> Title: `No information for this item`
#> Text:
#> I get `No information for this item` for a few of my series.
#> 
#> I managed to reproduce the bug:
#> 
#> 1) Create a new SA-Item with specification RSA3
#> 2) Then change the `Series span` to `From` with the date beginning the series
#> 
#> Surprisingly, the bug no longer appears when the series span is changed several times.
#> 
#> Here is an
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#821 ──
#> 
#> Title: Accepting new providers
#> Text:
#> With the emergence of new file formats, would it be costly to add new providers?
#> 
#> At Insee, we're thinking in particular of the parquet format (which is not currently a priority for time series data, but could become one in the future). 
#> 
#> ── Issue jdemetra/jdplus-main#789 ──
#> 
#> Title: Create td regressors based on custom groups
#> Text:
#> It would be great to be able to customize the calendar regressor sets by adding a national calendar and choosing our day groups by hands (in the GUI or in R).
#> 
#> By the way, what is the intention of the `TDuser` option in the GUI?
#> 
#> <img width="535" height="705" alt="Image" src="https://github.com/user-attachments/assets/
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#788 ──
#> 
#> Title: New TradingDays default groups
#> Text:
#> There are currently 7 different trading day configurations in JDemetra+ (GUI and R packages):
#> 
#> - `TD_NONE`
#> - `TD2` : mon-fri + sat-sun
#> - `TD2C` : mon-sat + sun
#> - `TD3` : mon-fri + sat + sun
#> - `TD3C` : mon-thu + fri-sat + sun
#> - `TD4` : mon-thu + fri + sat + sun
#> - `TD7` : all different days
#> 
#> At Insee, we are using the `T
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#773 ──
#> 
#> Title: Long-term mean with recurring `single_day`?
#> Text:
#> Recently, we encountered a situation where a regular public holiday (which occurs every year but is a moving holiday) was not included in the default list in JDemetra+ (see #766). 
#> 
#> So we manually added the corresponding days for several years using `single_day`.
#> 
#> In such a case (with `single_day` instead of `special_d
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#766 ──
#> 
#> Title: Islamic calendar
#> Text:
#> Currently, in the list of special days presented in JDemetra+ for the trading days correction, there are no day referring to the [Islamic holidays](https://en.wikipedia.org/wiki/Islamic_holidays).
#> 
#> Would it be expensive to implement them and offer them in JDemetra+?
#> 
#> Thank you.
#> 
#> ── Issue jdemetra/jdplus-main#758 ──
#> 
#> Title: Unify period display in GUI (2)
#> Text:
#> Issue #489 is unresolved. 
#> The problem remains in version 3.5.1. 
#> 
#> Also, the SI ratio display is still different and appears with Q1, Q2, Q3, Q4 :
#> 
#> <img width="605" height="255" alt="Image" src="https://github.com/user-attachments/assets/49948a60-a027-4bed-8947-9e9eedc49537" />
#> 
#> ── Issue jdemetra/jdplus-main#751 ──
#> 
#> Title: Structure of a WS
#> Text:
#> Hello, 
#> 
#> I'm opening an issue but it's more of **a discussion**.
#> 
#> First of all, the WS structure contains:
#> 
#> - data (calendar, regressors and data), 
#> - statistical models (specifications) 
#> - and outputs. 
#> 
#> In the context of moving to a cloud-native solution and changing JD+ to a web API, would it make sense to review th
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#733 ──
#> 
#> Title: Changing the `domainSpec` doesn't refresh the model
#> Text:
#> ### Intro
#> 
#> In the latest v3.5.1, there is a *breaking change* with v3.5.0 regarding the specification handling.
#> 
#> ### Step to reproduce
#> 
#> 1) Open JDemetra+ v3.5.1
#> 2) Create a new WS
#> 3) Add a series (with for example RSA3 specification)
#> 4) Change the `domainSpec` (**Right click** > **Edit** > **Reference specification** >
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#695 ──
#> 
#> Title: Infinity result for diagnostics.fcast-outsample-variance causes remainder of dem
#> Text:
#> I've found a bug with the cruncher. 
#> 
#> Here's the output I saw at the end of my run: 
#> - At the end of the output printed out by cruncher was this: (note the 'failed' line)
#> ```
#> Flushing bundle...
#> Generate Output
#> Csv matrix
#> failed: Input length = 1
#> csv
#> generated
#> Saving new processing...
#> ```
#> - output series were produced a
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#690 ──
#> 
#> Title: Change of reference specification in SAProcessing is not working
#> Text:
#> I changed the reference specification from RSAfull to x11 and then the specification wasn't changed:
#> 
#> <img width="991" height="603" alt="Image" src="https://github.com/user-attachments/assets/719f1e55-d3f6-4c3e-9d54-6ac304e7a84e" />
#> 
#> ── Issue jdemetra/jdplus-main#664 ──
#> 
#> Title: Combined seasonality tests on sa not appearing
#> Text:
#> Hi,
#> 
#> When I set my (monthly) series and model span to only the last 36, 37 or 38 periods, the 'Combined seasonality tests on sa' and 'Combined seasonality tests on sa (last years)' don't appear in the Main results part of the GUI:
#> 
#> Main results with longer series and model spans:
#> ![Image](https://github.com/user-attach
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#653 ──
#> 
#> Title: Excel SA-Item names contain `\n` to separate sheet series
#> Text:
#> When we create a new SA-Item from an Excel file, the name of the series is automatically `sheet name\nseries name`.
#> 
#> On the GUI, it appears with an arrow ▶:
#> 
#> ![Image](https://github.com/user-attachments/assets/955bd1e0-98aa-4101-bc25-8af37974715b)
#> 
#> But on the XML file:
#> 
#> ![Image](https://github.com/user-attachments/asse
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#640 ──
#> 
#> Title: Role of the metadata in the WS structure
#> Text:
#> Currently in a SA-Processing there are SA-Item and each SA-Item can contain a `<item name="metadata">` section with (currently only) the `<item name="comment">` item.
#> I understand that the role of this section is to add information that are not related to the adjustment but we want to keep.
#> 
#> But could we include in thi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#636 ──
#> 
#> Title: Error with `mvn clean install -DskipTests`
#> Text:
#> I try to build the jdplus_main project with `mvn clean install -DskipTests`, but I get an error:
#> 
#> ```
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD FAILURE
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time:  07:58 min
#> [INFO] 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#572 ──
#> 
#> Title: Cannot rename user-defined variables (regressors) successfully
#> Text:
#> Consider creating a Variables document with a regressor called 'x_1':
#> 
#> ![Image](https://github.com/user-attachments/assets/78726fbe-bfa3-4081-b8cb-43423b67eba2)
#> 
#> Now consider a series in a multi-document and applying the above regressor to it:
#> 
#> ![Image](https://github.com/user-attachments/assets/b39d481f-62e7-4f93-92fa
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#564 ──
#> 
#> Title: Missing item in dictionaries
#> Text:
#> I don't find some items in the dictionaries:
#> 
#> - in [RegressionDictionaries](https://github.com/jdemetra/jdplus-main/blob/develop/jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/main/java/jdplus/toolkit/base/api/dictionaries/RegressionDictionaries.java):
#>   - `regression.description`
#>   - `regressi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#562 ──
#> 
#> Title: Model changed by Seats - output
#> Text:
#> JDemetra 3.4.0 - When the TRAMO model is not decomposed, the representation (P,D,Q)(BP,BD,BQ)s  of the Seats model is not available.
#> 
#> ── Issue jdemetra/jdplus-main#557 ──
#> 
#> Title: Documentation for number pattern
#> Text:
#> Could it be possible to add some documentation to choose the **locale** corresponding to a `number pattern`:
#> 
#> ![Image](https://github.com/user-attachments/assets/0ae53d62-19d8-48a9-935f-37bdb174fde0)
#> 
#> Maybe a link or a small comment?
#> 
#> ── Issue jdemetra/jdplus-main#534 ──
#> 
#> Title: Checking for invalid series in workspace files
#> Text:
#> When the JD+ GUI indicates that a series has a 'Status' of 'Invalid', how is this recorded in the underlying XML workspace file?
#> 
#> ![Image](https://github.com/user-attachments/assets/5e230a9f-148a-4a27-8a60-a83b3c654e90)
#> 
#> Is the 'Quality' tag changed to 'Undefined'? Is such a change sufficient, or is there some other in
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#532 ──
#> 
#> Title: Unnecessary row displayed in Grid of time series
#> Text:
#> The following series has been imported from an Excel spreadsheet via the Providers panel. The series spans April 1982 to December 1985. However, when viewing the series in a Grid by double clicking on it, a row for 1986 is unnecessarily shown:
#> 
#> ![Image](https://github.com/user-attachments/assets/368a03b5-4e6f-44e1-bfeb
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#523 ──
#> 
#> Title: Add Copy Path/Reference context menu item to the Providers tab
#> Text:
#> The SDMX plugin uses a Copy Path/Reference context menu item to allow the user to quickly copy and paste relevant data IDs such as source, flow, and key. It is also used to create a command line with the appropriate parameters, allowing for some automation.
#> 
#> ![Image](https://github.com/user-attachments/assets/63c3066f-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#492 ──
#> 
#> Title: V3 output dictionary for Cruncher parameter file
#> Text:
#> Regarding the output dictionary at: https://github.com/jdemetra/jwsacruncher/wiki/Output-dictionaries, this appears to be for JD+ version 2. If that is the case, could we please kindly request an exhaustive list for version 3?
#> 
#> Please also see the following separate but related issues (especially the first one):
#> - h
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#485 ──
#> 
#> Title: Series vs items in the Cruncher input parameter file
#> Text:
#> Regarding the output dictionary for the input parameter file for Cruncher (https://github.com/jdemetra/jwsacruncher/wiki/Output-dictionaries), could a column please be added to clarify which keys refer to series as opposed to items?
#> 
#> That is, which tags correspond to the `<matrix>` items block, and which tags corresp
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#483 ──
#> 
#> Title: Precedence of refresh policy specified in Cruncher parameter file vs workspace f
#> Text:
#> When supplying the input parameter file to a Cruncher run as documented at https://github.com/jdemetra/jwsacruncher/wiki#parameters-file, the refresh policy must be specified:
#> 
#> ![image](https://github.com/user-attachments/assets/7f0546af-a2f8-4b4a-8120-a538393f4f88)
#> 
#> However, a refresh policy can also be specified 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#458 ──
#> 
#> Title: Make release process more robust against missing binaries in java distribution
#> Text:
#> Some builds fail because the java distribution bundled in JD+ (temurin) has sometimes some issues with the windows binaries.
#> The release process should be more robust to handle this kind of event.
#> As an alternative, JD+ could also use Azul Zulu instead of Adoptium Temurin.
#> 
#> Current workaround: https://github.com/jd
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#455 ──
#> 
#> Title: Manual creation of user defined variables
#> Text:
#> Hey development team,
#> I am trying to directly create a workspace for JDemetra using its Java libraries (e.g. jdplus-sa-base-workspace). I want to add Variables and userDefinedVariables.
#> When creating the Variables using the JD GUI, the variables show up in the XML file as dynamicTsVariable. One can add multiple of th
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#428 ──
#> 
#> Title: Truncated text in global SA general options tab
#> Text:
#> Observe the truncated text in the General tab of the SA options window highlighted in the screen shot below. Despite sufficient room for the text to be displayed in full, it has been truncated. (Presumably, it should say 'Model stability'.)
#> 
#> ![image](https://github.com/user-attachments/assets/20dcec28-c637-4895-bd9b-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#423 ──
#> 
#> Title: Documentation on workspace file XML tags (seasonal adjustment parameter settings
#> Text:
#> Is there documentation on all possible XML tags (that is, all seasonal adjustment parameter settings as per the Specifications panel) that could possibly populate a workspace file using the X13 algorithm?
#> 
#> If not, could we please request such documentation, specifically doco that contains:
#> - all possible XML tags/pa
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#421 ──
#> 
#> Title: Missing revisions policy in documentation
#> Text:
#> According to the list of revisions policies for JDemetra+ at: https://github.com/jdemetra/jwsacruncher/wiki#parameters-file, there are 8 of them:
#> 
#> ![image](https://github.com/user-attachments/assets/807fdf98-3a8f-4fe0-b9fd-c5e3dd3d2eb7)
#> 
#> However, at: https://jdemetradocumentation.github.io/JDemetra-documentation/pa
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#399 ──
#> 
#> Title: How to change the 'bias' parameter in the GUI?
#> Text:
#> The XML file underlying each workspace includes the following `bias` tag by default:
#> 
#> ![image](https://github.com/user-attachments/assets/7d0bda1a-030e-4078-a06a-5041d2ca9c1f)
#> 
#> Two questions:
#> 
#> - What is it? (What does it mean or do, and what values can it take?)
#> - Where can it be viewed, changed or set in the Sp
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#381 ──
#> 
#> Title: Allow separation of excluding forecasts vs backcasts in the generation of extrem
#> Text:
#> A potential enhancement of the 'Excludeforecast' parameter is to separate it into excluding forecasts and excluding backcasts individually, so that forecasts can be excluded in the generation of extreme values whilst backcasts are not excluded, or vice versa.
#> 
#> ![image](https://github.com/user-attachments/assets/54657
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#372 ──
#> 
#> Title: The name of the newly created WS does not appear in the GUI
#> Text:
#> When a ws is created and saved in version 3, its name does not appear in the bar provided.
#> To make it appear, close the ws then reopen it and the name will be displayed correctly.
#> 
#> Here is a small demo of what I did to get this bug:
#> 
#> https://github.com/jdemetra/jdplus-main/assets/112711017/963115ec-a1ac-421f-b237-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#345 ──
#> 
#> Title: Format type of date in output
#> Text:
#> Dear all, 
#> 
#> I noticed some trouble with the new output format: 
#> in version 2, the date was of format "General" in the output, however 
#> in version 3, the date is changed to a real date of format type: 
#> 1995-01-01T00:00:00/PM3, which creates problems in our subsequent macros. 
#> Is there any special reason behind thi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#338 ──
#> 
#> Title: Restoring series name in legend does nothing
#> Text:
#> When I rename a series in the legend of a chart and click on 'Restore', nothing seems to happen. For example, in the image below, the initial names of the series that I used to create the chart were both 'Seasonally adjusted'. I changed this to 'SA (with auto-outliering)' and 'SA (without auto-outliering)'. When I then
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#335 ──
#> 
#> Title: Open JDemetra+ workspace with `Open with...`
#> Text:
#> When I want to open an existing workspace with JDemetra+, I open the GUI then I go to `File` --> `Open Workspace` --> then I search my `.xml` file.
#> 
#> Is it possible to open a workspace (from the `.xml` file) with the `Open with...` button?
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/112711017/d21266de-7
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#315 ──
#> 
#> Title: Missing diagnostics in Summary
#> Text:
#> In the main panel of a SA-Item (**Main Results**), there are sometimes some missing diagnostics in the **Residuals seasonality tests**.
#> 
#> A classic series contains 4 items in the **Residuals seasonality tests** : two QS-Test and two F-test:
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/112711017/9dafb92b-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#283 ──
#> 
#> Title: Plot and display series if SA fail (and additionnal explanations)
#> Text:
#> (Issue related to jdemetra/jdemetra-core#1129)
#> 
#> If the seasonnal adjustment fails, I can't neither plot nor see my values.
#> 
#> For example here, I have a dataset with 11 values so the seasonal adjustment will obviously fail:
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/112711017/818fc2b7-9885-4627-8edc-b
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#270 ──
#> 
#> Title: warning at start of 3.2.2
#> Text:
#> warning message when starting 3.2.2
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/43274870/e4291603-3036-4f16-a1b5-a42dd360af11)
#> 
#> Details 
#> org.netbeans.core.network.proxy.pac.PacParsingException: Cannot find secure PAC script engine.
#> Allowed engines: GraalVM:js,Graal.js,Nashorn
#> Found engines:
#> Will no
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#261 ──
#> 
#> Title: No new logo in shortcuts
#> Text:
#> the new logo appears just in the top left corner of the open GUI but not in the short cuts 
#> 
#> 
#> ![image](https://github.com/jdemetra/jdplus-main/assets/43274870/d5c4ca0c-9d70-42ee-bd40-97a2714db224)
#> 
#> 
#> ── Issue jdemetra/jdplus-main#250 ──
#> 
#> Title: Add time filter in UI components
#> Text:
#> See https://github.com/jdemetra/jdplus-main/blob/develop/jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-api/src/main/java/jdplus/toolkit/base/api/timeseries/TimeSelector.java
#> 
#> ── Issue jdemetra/jdplus-main#177 ──
#> 
#> Title: Improve branding
#> Text:
#> Current branding of JD+ is outdated and inconsistent.
#> 
#> Here is a non-exhaustive list of items to update and improve:
#> 1. icon, favicon and avatar (square shape, logo only)
#> 2. splash screen, about dialog image and banner (rectangular shape, landscape orientation, logo+text)
#> 3. short description (a few words)
#> 4. lon
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#161 ──
#> 
#> Title: Use enhanced Pseudo-Random Number Generators from JDK 17
#> Text:
#> See https://openjdk.org/jeps/356
#> 
#> ── Issue jdemetra/jdplus-main#136 ──
#> 
#> Title: [FR] decimal separator in the JWSACruncher
#> Text:
#> It could be good to be able to change the decimal separator in the JWSACruncher to avoid a system dependency. This is linked to https://github.com/jdemetra/jwsacruncher/issues/13.
#> 
#> ── Issue jdemetra/jdplus-main#112 ──
#> 
#> Title: JWSACruncher and parameters `fullseriesname` and `refreshall`
#> Text:
#> I tried the following parameters files an a workspace with data created by an Excel file and it has no effect. Are the parameters `fullseriesname` and `refreshall` suppose to work in the JWSACruncher? Am I doing something wrong?
#> ```
#> <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
#> <wsaConfig bundle="10000" cs
#> ...
#> 
with_comments(all_issues, negate = TRUE)
#> There are 70 issues. 
#> 
#> ── Issue jdemetra/jdplus-main#1029 ──
#> 
#> Title: Seasonal component in X11 spec : if set to NO, no info displayed in GUI
#> Text:
#> HI ,
#> 
#> in v 3.8.0 if seasonal component is unticked in the X11 part of the spec 
#> 
#> No information is diplayed in GUI  
#> 
#> <img width="246" height="62" alt="Image" src="https://github.com/user-attachments/assets/a57354d6-256f-4bac-aff8-fa68720aee74" />
#> 
#> the estimation appears as invalid 
#> 
#> <img width="647" height="40" alt="I
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#1006 ──
#> 
#> Title: Benchmarking Results includes forecast observations beyond the input series
#> Text:
#> <img width="340" height="216" alt="Image" src="https://github.com/user-attachments/assets/cd77b619-d644-4d03-925e-6a858eeb3709" />
#> 
#> While testing JDemetra+ 3.7.1, I noticed an unexpected behaviour in the **Benchmarking Results** output. The input series used in this example ends in **2025 Q1**. However, the table displ
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#1004 ──
#> 
#> Title: Outlier type and position are missing in Matrix |> Outliers
#> Text:
#> Related to cosa-tsa-shop/UserSupport#10
#> 
#> In the section Matrix |> Outliers (from a SA-Processing), the version 2 and 3 differs by their presentation:
#> 
#> - In version 2 we observe a column by outliers with type, position, coeff and T-Stats
#> - In version 3 we observe 2 columns by outliers with coeffs and T-Stats
#> 
#> Is it poss
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#962 ──
#> 
#> Title: JDemetra+ doesn't import duplicated variables
#> Text:
#> If two variables have the same value, only the last one will be read.
#> 
#> Here is an example in R:
#> 
#> ``` r
#> library("rjd3workspace")
#> path_xml <- tempfile(fileext = ".xml")
#> writeLines(
#>     text = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
#> <tsVariables xmlns="ec/tss.core">
#>     <tsVariable name="VAR1">
#>         <t
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#943 ──
#> 
#> Title: Generating output only if the SAP is computed
#> Text:
#> I try to understand why the output generation wors with the GUI sometimes and sometimes not.
#> 
#> In this issue, I will use the verbs "compute" and "estimate" to refer to the same thing: obtaining the results of the SA estimate for my series. But perhaps there is a fundamental difference between the two actions...
#> 
#> ## Exam
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#940 ──
#> 
#> Title: How the GUI sort by specification?
#> Text:
#> In the GUI, it's possible to sort the SA-Item of a SA-Processing by **Reference spec**. But how is it handle?
#> 
#> I notice (in this example) that this isn’t done by name and that the `^` and `v` keys aren’t used to reverse the direction.
#> 
#> https://github.com/user-attachments/assets/c627e166-e28b-4de2-818f-821dd96c6b6c
#> 
#> ── Issue jdemetra/jdplus-main#888 ──
#> 
#> Title: Is it possible to choose the value `(?)` in GUI output?
#> Text:
#> The aim of this issue is to find out whether, when I generate an excel output with the GUI, it's possible to choose the length of series which items ends with `(?)`.
#> 
#> I will take the example of the forecasted series. 
#> 
#> In X13-RegARIMA, there are forecasts in the preprocessing part but only with `(?)` items and it's not
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#883 ──
#> 
#> Title: Adding a button to "refresh" a WS
#> Text:
#> *related to #882*
#> 
#> I'm **not** talking here about the refresh policy that can be applied to a SA-Item or to a SA-Processing. 
#> I'm talking about the action that is executed each time a workspace is saved or when it is closed and reopened. The WS is saved in a **XML format** but some information are also updated (maybe i
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#871 ──
#> 
#> Title: No item with `(?)` with the GUI
#> Text:
#> Related to #594 
#> 
#> The GUI output does not allow you to specify the number of elements to be displayed in the boxes `(?)`.
#> 
#> For certain variables, it is interesting to request additional columns (in particular, diagnostics, modality, comments, values and p-values...). For cruncher queries, you can provide the number of 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#870 ──
#> 
#> Title: The GUI generate no output for residuals
#> Text:
#> For some items (of the output list), the GUI doesn't produce any output.
#> Here I tested with the items `residuals.skewness`, `residuals.kurtosis`, `residuals.doornikhansen`, `residuals.lb`, `residuals.bp` , `residuals.lb2`.
#> 
#> https://github.com/user-attachments/assets/0569371a-6617-4da9-843e-b14ffdadee88
#> 
#> 
#> Here is the [w
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#869 ──
#> 
#> Title: Free space hides elements
#> Text:
#> When certain displays open (pop-up boxes, panels, tabs, etc.), certain banners are hidden.
#> 
#> For example, when the output generation dialog box opens, the height of the red and green boxes is preset. When I click on `Full series name`, the description doesn't appear completely (in the green box):
#> 
#> <img width="578" heigh
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#866 ──
#> 
#> Title: Declare and create relative paths with R
#> Text:
#> It would be really nice to be able to create relative paths from the cruncher (with the command line) or via R packages.
#> 
#> The aim would be to be able to create `Demetra Paths` or `Favorites`, as in the GUI.
#> 
#> An example would be to define paths paths that would apply directly à my WS (or à all JDemetra+ sessions by writ
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#864 ──
#> 
#> Title: Add seed to random function
#> Text:
#> Is it possible to add a seed parameter to method that generate random number?
#> 
#> It concerns the functions:
#> 
#> - [`randomsT`](https://github.com/jdemetra/jdplus-main/blob/88cb98a6f80adbbe3566f31e8ed910f1e46f1d36/jdplus-main-base/jdplus-toolkit-base-parent/jdplus-toolkit-base-r/src/main/java/jdplus/toolkit/base/r/stats/Dist
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#861 ──
#> 
#> Title: Add ability to close a workspace
#> Text:
#> It might be an idea to add `Close Workspace` to the File menu in the JD+ GUI:
#> 
#> <img width="219" height="281" alt="Image" src="https://github.com/user-attachments/assets/8744ae6c-fb9f-466b-8804-fc70b686cfea" />
#> 
#> I was looking for this feature one time when I wanted to reopen an existing workspace and lose my changes (as
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#847 ──
#> 
#> Title: Cannot press Esc to close certain dialogue boxes
#> Text:
#> Consider opening a new workspace via **File | Open Workspace** and having the corresponding dialogue box appear. If the user presses the `Esc` key on the keyboard, then it will conveniently close, as expected.
#> 
#> However, dialogue boxes like the 'Pre-specified outliers' or 'Variables' ones (opened via the Specifications 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#843 ──
#> 
#> Title: Why the color is not updated?
#> Text:
#> ## Questions
#> 
#> ### Not the color
#> 
#> When I try to switch from a Tramo-Seats specification to X13 (or vice versa), the `domainSpec` (**Reference Specification**) name is updated, but **not the color**.
#> 
#> ### Apply before refresh
#> 
#> Then, if I want to refresh the model, I have to **apply the `domainSpec`** before I can use a r
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#842 ──
#> 
#> Title: Zoom in GUI
#> Text:
#> It would be great to be able to zoom in the GUI.
#> 
#> For example, here is a normal view with big screen:
#> 
#> <img width="1920" height="1032" alt="Image" src="https://github.com/user-attachments/assets/e9f0a942-daa0-41b1-8980-09823cda08a9" />
#> 
#> We could zoom in = increase the size of the fonts, the width, the height of the con
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#841 ──
#> 
#> Title: Overview of a theme
#> Text:
#> ## Questions
#> 
#> ### Change and preview a theme
#> 
#> When you want to change the theme in the GUI, you have to restart the interface. It would be interesting to have a preview (without necessarily updating the interface) to choose a theme.
#> 
#> ### Light and dark
#> 
#> Also, in many programs, it is possible to choose a “light” and a “
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#838 ──
#> 
#> Title: Cannot delete several objects at the same time
#> Text:
#> For certain types of object, it is not possible to delete them all at the same time.
#> 
#> These include:
#> 
#> - Several groups of variables:
#> 
#> <img width="310" height="226" alt="Image" src="https://github.com/user-attachments/assets/63bc2edf-a114-4a60-a230-485f7fb92808" />
#> 
#> - Several SA-Processings:
#> 
#> <img width="313" height="14
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#837 ──
#> 
#> Title: `Delete`, `Close`, `Clear` or `Remove`
#> Text:
#> The vocabulary used to remove or delete an element is not the same:
#> 
#> - **Remove** is used for variables
#> - **Delete** is used for SA-objects (SAI, SAP, SAD and WS)
#> - **Close** is used for provider connections
#> - **Clear** for containers (grid, group of regressors and chart)
#> 
#> 
#> ## Examples
#> 
#> ### `Remove`
#> 
#> Removing variables
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#819 ──
#> 
#> Title: Regarima residuals threshold values
#> Text:
#> We have noticed that threshold values in `Tools > Options > SA > TramoSeats > Regarima residuals` are the same for Uncertain/Bad for Seasonal and Severe/Bad for Trading days (see picture below).
#> When changing the settings and closing the program, the changes are not saved, they default to the same value.
#> 
#> <img width="4
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#816 ──
#> 
#> Title: Refreshing a csv file without reload does not work
#> Text:
#> There is 2 part to this issue.
#> 
#> ## Reloading automatic
#> 
#> I'm working with a csv file (open with Notepad++ in the video) and a .xlsx file (open with LibreOffice in the video).
#> 
#> When I update datasets with a new point, the spreadsheet is automatically updated while the csv file needs to be reloaded:
#> 
#> https://github.com/us
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#809 ──
#> 
#> Title: Improve JExceptionPanel to make it more readable for regular users
#> Text:
#> The current JExceptionPanel displays an exception as a stacktrace text with some fancy colors.
#> 
#> This component could be improved to
#> - automatically unwrap exceptions to unclutter the text
#> - use a JTree for the rendering so nodes could be hidden
#> - add a search function
#> - link to a db of common problems
#> - highlight URIs,
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#798 ──
#> 
#> Title: Chart container legend labels shifting undesirably
#> Text:
#> Consider creating a chart and superimposing a number of chronologically based series on to it by using the freezing functionality.
#> 
#> Unfortunately, upon right clicking on a series in the chart and selecting 'Freeze', the legend label for that series is shifted to the right-most spot (and appended with the string: '(froz
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#774 ──
#> 
#> Title: Access the path to the raw data underlying a SA-Item
#> Text:
#> A great new functionality would be to be able to access the path to the raw data underlying an SA item and be able to open the file or at least the folder containing the file.
#> For example, right-click on an SA item -> ‘Open raw data’.
#> 
#> <img width="597" height="385" alt="Image" src="https://github.com/user-attachments/a
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#772 ──
#> 
#> Title: With which calendar and regressors are computed the residual trading days tests
#> Text:
#> Duplicated of jdemetra/jdemetra-app#373 in v3.
#> 
#> The question is the same: 
#> 
#> 1) On what calendar basis are residual trading days tests calculated?
#> 2) Is it possible to use a custom calendar to compute these tests?
#> 3) What happens if there are multiple calendars in **Calendars**?
#> 
#> 
#> ── Issue jdemetra/jdplus-main#768 ──
#> 
#> Title: UX difficult to use sometimes
#> Text:
#> related to #349 
#> 
#> Sometimes, the interface is difficult to use and not intuitive.
#> 
#> The list includes:
#> 
#> - [ ] Buttons in ‘empty space’ where you have to click next to the item
#> - [ ] The drop-down buttons (`>`) for submenus are sometimes a little small (or unintuitive)
#> - [ ] Clicking ON the item and not next to it to sel
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#767 ──
#> 
#> Title: Holidays to choose the calendar
#> Text:
#> In the GUI, the button to add a *custom calendar* to compute the trading days regressor is called **holidays**. 
#> 
#> Should it be renamed **calendar** ?
#> 
#> 
#> <img width="305" height="462" alt="Image" src="https://github.com/user-attachments/assets/af66f4db-1f86-40cf-800f-c46d3effdcb5" />
#> 
#> ── Issue jdemetra/jdplus-main#763 ──
#> 
#> Title: Error loading V2 workspace with fixed week calendar day
#> Text:
#> If I create a workspace in V2 with a National Calendar with a Fixed Week it cannot be imported with the V3. Error loading the calendar with the following error:
#> ```
#> java.lang.NullPointerException: Cannot invoke "String.toUpperCase(java.util.Locale)" because "day" is null
#> 	at internal.toolkit.base.workspace.file.xml.uti
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#752 ──
#> 
#> Title: Add ways to retrieve WS file names and paths
#> Text:
#> It would be very interesting to add ways to retrieve WS file names and paths.
#> 
#> Is it possible to add methods like `getSource` and `setSource` to be able to retrieve the file containing a SA-Processing, a SA-Item or even a Workspace directly from the Java object?
#> 
#> Thank you
#> 
#> ── Issue jdemetra/jdplus-main#734 ──
#> 
#> Title: Can't refresh a SA-Item when WS newly created
#> Text:
#> ### Introduction
#> 
#> In JDemetra+ v3, it's not possible to refresh a SA-Item with **Right click** > **Refresh** when the newly created WS is not saved.
#> It was possible to refresh a new WS in version 2.
#> 
#> But there is A way to refresh a SA-Item with concurrent policy with the GUI in version 3.5.0 with a newly created WS is 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#698 ──
#> 
#> Title: Series naming restrictions
#> Text:
#> This is just a question:
#> 
#> Are there any character restrictions imposed by the JDemetra+ GUI and/or the Cruncher when naming series (that is, when providing a name or ID for series)? For example, are certain special characters not allowed, or are there any character-length limits, and do these differ for the GUI vs the 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#687 ──
#> 
#> Title: Output from Cruncher and Output from UI have a small inconsistency for fixed use
#> Text:
#> Please find enclosed the example for a user-defined trading day. The output could be generated with the cruncher or with output in the UI for a SaProcessing. 
#> In the output in "ahaCN_GP14\SAProcessing-1\demetra_m.csv" for ahaCN_GP14, the regression.ntd is 1 but the column E and F for the variables are empty. In Version
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#682 ──
#> 
#> Title: Make outlier information consistent in Pre-processing panel for fixed and non-fi
#> Text:
#> All fixed and non-fixed (estimated) outlier corrections can be conveniently seen in the Pre-processing panel. However, the information displayed for each type (fixed vs non-fixed) differs. In the example screen shot below, non-fixed (pre-specified) outliers have the type and period shown, but not the name; whereas for 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#679 ──
#> 
#> Title: Cruncher automatically changing transform function
#> Text:
#> Hi,
#> 
#> We have noticed that when a series that has regarima.transform.function = 'Log' cannot be logged (for example, if it has 0s or negative values in the original series), the Cruncher automatically converts the transform function to 'None'.
#> 
#> Are you able to confirm that this is the only situation in which the Crunche
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#652 ──
#> 
#> Title: The cruncher generates no warning if failing to compute
#> Text:
#> We can study 3 different cases of data:
#> 
#> - ts without metadata
#> - ts with metadata
#> - ts with false metadata
#> 
#> In these 3 cases, the seasonal adjustment works, but the refresh only works for the first 2 cases and generates an invalid for the last case.
#> 
#> **Is this a deliberate behaviour?**
#> 
#> here is my R code:
#> 
#> ```r
#> library
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#651 ──
#> 
#> Title: How do you re-order SAIs in SAP?
#> Text:
#> Is it possible to move (up or down) an SA-Item in an SA-processing?
#> 
#> I'd like to reorganise my SA-Items so that they follow a certain nomenclature (a certain order).
#> I'd like to do this in R but I haven't found any other method than creating a copy of my WS and transferring my SA-Items one by one in the order I want.
#> 
#> 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#645 ──
#> 
#> Title: `No information for this item` and `Invalid` when Variables have spaces
#> Text:
#> When you create a WS with variables and the variable name contains one or more spaces, JDemetra+ no longer knows how to distinguish variables from their name and you end up with more variables than you expected and incorrect variables.
#> 
#> Here's an example of my tests on video:
#> 
#> https://github.com/user-attachments/assets
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#643 ──
#> 
#> Title: Several synonyms and imprecisions in the JDemetra+ term
#> Text:
#> In JDemetra+, there is a precise vocabulary for distinguishing objects.
#> **Names are important** and some terms are unfortunately synonyms and can refer to the same object.
#> 
#> I suggest grouping the names together to obtain one name = one object.
#> 
#> Here's my list:
#> | Current name                                             
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#642 ──
#> 
#> Title: Error with `mvn clean install`
#> Text:
#> I try to build the jdplus_main project with `mvn clean install`, but I get an error:
#> 
#> ```
#> [INFO] ------------------------------------------------------------------------
#> [INFO] BUILD FAILURE
#> [INFO] ------------------------------------------------------------------------
#> [INFO] Total time:  01:38 h
#> [INFO] Finished at: 2
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#608 ──
#> 
#> Title: Add export to mermaid.js chart
#> Text:
#> Description from Mermaid.js :
#> > JavaScript based diagramming and charting tool that renders Markdown-inspired text definitions to create and modify diagrams dynamically.
#> 
#> Example :
#> 
#> This text ...
#> ```
#> xychart-beta
#>     title "Sales Revenue"
#>     x-axis [jan, feb, mar, apr, may, jun, jul, aug, sep, oct, nov, dec]
#>     y-axi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#603 ──
#> 
#> Title: Cannot edit format of charts
#> Text:
#> When I right-click on a chart container and select **Edit format...** to change the formatting of the dates and numbers (such as using the settings below), nothing changes:
#> 
#> ![Image](https://github.com/user-attachments/assets/60d7c8fb-8724-461c-9fe0-3664dbbf7da0)
#> 
#> ![Image](https://github.com/user-attachments/assets/315
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#596 ──
#> 
#> Title: Add OpenNodeAction as preferred action on ProviderNode when node has no children
#> Text:
#> A provider node has no children if no datasource has been defined and/or restored.
#> Double-clicking on it in this state doesn't do anything.
#> Currently, the user must right-click the provider node to add a new source.
#> The idea is to check if the provider node has no children on double-click and, if so, add a new source d
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#594 ──
#> 
#> Title: Missing series items in the output of the GUI
#> Text:
#> Related to #294 
#> 
#>  ### GUI v3 vs {rjd3x13}
#> 
#> Normally, the graphical interface and R packages in version 3 produce the same output. 
#> However, some items produce an output with **{rjd3x13}** that cannot be obtained via the GUI output:
#> 
#> - `decomposition.s_cmp_f`
#> - `decomposition.sa_cmp_f`
#> - `decomposition.t_cmp_f`
#> - `l`
#> -
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#593 ──
#> 
#> Title: Periodicity/frequency of series not immediately clear
#> Text:
#> For series in a multi-document, it's not immediately obvious to the user what the periodicity or frequency of the series is. After processing an individual series and clicking on its row, the 'Main results' tab appears and the user can tell the frequency by the x-axis labelling of the SI chart as highlighted in red bel
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#591 ──
#> 
#> Title: User-defined variables renaming is not consistent
#> Text:
#> When renaming User-defined variables through the following menu, I can sometimes rename variables to the same name, and other times I get an error message.
#> 
#> ![Image](https://github.com/user-attachments/assets/b2b30a7e-cc26-44c0-9b71-4fb3c7b8bb95)
#> 
#> Here, I attempted to rename Headline_impute_2qtrs to Headline imputation
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#590 ──
#> 
#> Title: Summary info on pre-specified outliers not clear
#> Text:
#> In the JD+ GUI, the user can insert 'pre-specified outliers' as shown below, and these can be fixed or non-fixed:
#> 
#> ![Image](https://github.com/user-attachments/assets/d4a65a9a-3173-4e67-80a3-1e1f4cdd7ee7)
#> 
#> If the user then looks at the **Summary** section of the **Main **results**** section in the output area for the s
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#571 ──
#> 
#> Title: List of user-defined variables (regressors) displayed poorly in 'Variables' dial
#> Text:
#> When I attempt to insert/modify/remove outliers via the 'Variables' dialogue box, the name of the variable that I am currently editing does not appear correctly. In the screen shot below, I have clicked the `+` button to insert a new variable, but it is barely visible on the left-hand side of the box. It seems to have 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#560 ──
#> 
#> Title: XML files are badly indented from a certain level upwards
#> Text:
#> Everything is in the title. From a certain level, the indentation of the XML files comes back to 0:
#> 
#> ![Image](https://github.com/user-attachments/assets/5eac083a-6ea7-4dcd-b271-47916001eb47)
#> 
#> This makes reading these files very complicated. Is it possible to indent these files correctly?
#> 
#> ── Issue jdemetra/jdplus-main#559 ──
#> 
#> Title: Default MCD in MCDFilter
#> Text:
#> When all the I/C ratios are greater than 1, the MCD value used in the program is equal to 2: 
#> https://github.com/jdemetra/jdplus-main/blob/37c578dbff38f6c54f682536222c185c718eb4d7/jdplus-main-base/jdplus-x13-base-parent/jdplus-x13-base-core/src/main/java/jdplus/x13/base/core/x11/filter/X11FilterFactory.java#L40-L60
#> Is 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#510 ──
#> 
#> Title: Series selection blocking highlighting of rows in check last test
#> Text:
#> Regarding the check last function in the JDemetra+ GUI, the red/orange highlighting in the 'Rel. Error N-1' column is blocked whenever that series' row is selected. In the example below, the series in the first row is selected, but the red highlighting in the relative error column cannot be seen:
#> 
#> ![image](https://gi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#491 ──
#> 
#> Title: Add confidence interval in the distribution chart of the residuals.
#> Text:
#> *Issue revealed by Nikolina Rizanovska*
#> 
#> Is it possible to add dashed lines to indicate the confidence interval at 95% in the distribution chart of the residuals?
#> (We are in the `Preprocessing` --> `Residuals` --> `Distribution` section)
#> 
#> ![image](https://github.com/user-attachments/assets/1ed8e0ef-bcf7-48cd-aac5-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#488 ──
#> 
#> Title: Should leap year be considered as trading day regressor?
#> Text:
#> *Issue revealed by Nikolina Rizanovska*
#> 
#> When you want to use calendar regressors (**including the leap year**) in your seasonal adjustment, JDemetra+ indicates a number of variable (here 6) in the Summary:
#> 
#> ![image](https://github.com/user-attachments/assets/e41846f1-dba3-4abe-8424-8a152a458e9f)
#> 
#> But it doesn't 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#484 ──
#> 
#> Title: Cruncher error logs
#> Text:
#> Hi JDemetra+ devs,
#> We are using cruncher v3 in a data processing application implemented in Java. At a high level we are calling jdplus.cruncher.App.main with a valid workspace and parameters file, then parsing the output files for consumption by downstream services.
#> One issue with cruncher is that in most cases it d
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#479 ──
#> 
#> Title: Possible mistake in documentation for unit root limit options
#> Text:
#> Regarding the entry for the unit root limit on this page: https://jdemetradocumentation.github.io/JDemetra-documentation/pages/reference-manual/modelling-spec-arima.html
#> 
#> ![image](https://github.com/user-attachments/assets/03153d3f-56d4-439a-83d5-fd363090fcb4)
#> 
#> In the underlying XML workspace file, the correspondin
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#449 ──
#> 
#> Title: demetra_m.csv empty column names
#> Text:
#> In the demetra_m.csv file, empty column names appear as in version 2. It would be more useful to have a standard name for each column.
#> ![demetram_empty_columnames](https://github.com/user-attachments/assets/f08575e5-ed8a-40df-8ca8-99c4af149dd1)
#> 
#> 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#447 ──
#> 
#> Title: user-defined calendar effect names do not appear in demetra-m.csv
#> Text:
#> user-defined calendar effect names do not appear in demetra-m.csv file.
#> ![calendar effec_demetram](https://github.com/user-attachments/assets/267217bf-e37c-40bd-913e-324a5dc48d79)
#> 
#> 
#> 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#429 ──
#> 
#> Title: Ability to specify non-standard parameters for prior corrections is absent in th
#> Text:
#> In RJDemetra, we are able to specify additional parameters for prior corrections as appropriate. For example, for level shifts (LSs), we can choose to have them as being 'zero-ended' or 'one-ended'. Is this possible in the GUI? There is no option in the relevant dialogue box for LSs, for example:
#> 
#> ![image](https://gi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#419 ──
#> 
#> Title: User-defined calendar effects
#> Text:
#> Better handling of user-defined calendar effects (output, gui)
#> 
#> ── Issue jdemetra/jdplus-main#418 ──
#> 
#> Title: Better handling of coefficients/parameters in the GUI
#> Text:
#> Improve the naming of coefficients/parameters in the spec boxes
#> Hide coefficients when they can't be specified (undefined transformation)
#> 
#> ── Issue jdemetra/jdplus-main#400 ──
#> 
#> Title: How to specify the data format in Providers?
#> Text:
#> When I try to add some series (txt files) in the Providers, by default there is a data format `1,234.5` (representing the value 1234.5):
#> 
#> ![image](https://github.com/user-attachments/assets/77695364-1999-468a-a066-502e323fb45a)
#> 
#> But how to change it with other formats as `1.234,5`.
#> There are a lot of different for
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#371 ──
#> 
#> Title: Add message in the case of redundant regression variables
#> Text:
#> It seems to me that in the presence of multicollinearity JD+ will declare one regression variable from each possible linear combination as being redundant and automatically set the corresponding regression effects to zero. The latter are then reported in a separate output table without any further explanation.
#> I think
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#353 ──
#> 
#> Title: Restore support of ODBC connection string on files 
#> Text:
#> In v2, it was possible to drag&drop files to the ODBC provider in order to create a connection string on the fly.
#> It was limited to MS Access files but could be extended to any file type supported by an ODBC driver.
#> 
#> ![Screenshot 2024-06-27 084947](https://github.com/jdemetra/jdplus-main/assets/8778378/2d58f93c-07f3
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#331 ──
#> 
#> Title: Improve SA output
#> Text:
#> Two problems with current implementation:
#> - the notification doesn't show the created file on Excel output
#> ![Screenshot 2024-06-14 133226](https://github.com/jdemetra/jdplus-main/assets/8778378/4664035b-69e5-4d18-a656-4f49a7104baa)
#> ![Screenshot 2024-06-14 133329](https://github.com/jdemetra/jdplus-main/assets/877837
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#296 ──
#> 
#> Title: Display jump discontinuity for level shifts (trend breaks)
#> Text:
#> It would be ideal to have the option available to view a jump discontinuity at any level shift (trend break) when plotting the trend. In the image below, for example, instead of the blue trend line continuing unbroken, there is a visual jump where the level shift occurs in Q4 2001. This makes is easier and quicker for 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#293 ──
#> 
#> Title: Add support of standard reporting period and time range
#> Text:
#> JD+ only supports dates (`2010-01-01`) as I/O for observation periods.
#> Other representations exists as well and could be integrated into a future release such as:
#> - standard reporting period (`2010-Q1`, `2012-M12`)
#> - time range (`2010-01-01/P3M`, `2012-12-01/P1M`)
#> 
#> These representations are already defined in the 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#287 ──
#> 
#> Title: Show internal corrections on SI chart
#> Text:
#> There are two potential improvements to SI charts in the JDemetra+ GUI:
#> 
#> - revealing X11 internal automatic corrections by simultaneously displaying modified and unmodified SI ratios with lines joining them
#> - (if not the above, or possibly in addition to:) allowing the user to choose between displaying modified or u
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#243 ──
#> 
#> Title: Color scheme selection via 'Configure...' not user-friendly
#> Text:
#> (Version: 3.2.2-SNAPSHOT)
#> 
#> - A graph's display options can be modified via the 'Configure...' context menu when right-clicking on it. This brings up a dialogue box with an option to change the 'Color scheme'. However, the different colour options are not presented as a user-friendly drop-down list. Instead, the user 
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#162 ──
#> 
#> Title: Use Platform-Specific Desktop Features from JDK 9
#> Text:
#> See https://openjdk.org/jeps/272
#> 
#> Most [nbbrd desktop API](https://github.com/nbbrd/java-desktop-util/) can be replaced by the latest JDK desktop API. 
#> There are a few exceptions anyway: search and kwown folders
#> 
#> ── Issue jdemetra/jdplus-main#130 ──
#> 
#> Title: XSD avaliability
#> Text:
#> As our organization are moving towards XMLs, it is getting too time consuming to adopt the XML files to the structure that JDEMETRA+ requries for input. 
#> 
#> Have this very easy code snippet to do it, given you have pivoted it in a cetrain way, but should be covered by someone who has far better knowledge. 
#> 
#> For bette
#> ...
#>