Check for text in GitHub Issues

Generic function to filter issues with a given text pattern in the title, body, or comments of a GitHub Issue object or a collection of Issues.

with_text(x, ...)

# S3 method for class 'IssuesTB'
with_text(x, ..., in_title = TRUE, in_body = TRUE, in_comments = TRUE)

Arguments

x

An object of class IssuesTB.

...

Additional arguments passed to grepl(), such as pattern and ignore.case.

in_title

Boolean. Does the function search for text in the title? (Default TRUE)

in_body

Boolean. Does the function search for text in the body? (Default TRUE)

in_comments

Boolean. Does the function search for text in the comments? (Default TRUE)

Value

An object IssuesTB with issues that satisfy the condition.

Examples

all_issues <- get_issues(
    source = "local",
    dataset_dir = system.file("data_issues", package = "IssueTrackeR"),
    dataset_name = "open_issues.yaml"
)
#> Looking into open_issues.yaml ...
#> The issues will be read from /home/runner/work/_temp/Library/IssueTrackeR/data_issues/open_issues.yaml.
with_text(all_issues, pattern = "Excel")
#> There are 8 issues. 
#> 
#> 
#> ── Issue jdemetra/jdplus-main#888 ──
#> 
#> Title: Is it possible to choose the value `(?)` in GUI output?
#> Text:
#> The aim of this issue is to find out whether, when I generate an excel output with the GUI, it's possible to choose the length of series which items ends with `(?)`.
#> 
#> I will take the example of the forecasted series. 
#> 
#> In X13-RegARIMA, there are forecasts in the preprocessing part but only with `(?)` items and it's not
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#865 ──
#> 
#> Title: `(frozen)` in SAI names
#> Text:
#> I just created a new WS with csv and Excel files. 
#> When I try to read it from R, I get `(frozen)` in the name of the SA-Items:
#> 
#> https://github.com/user-attachments/assets/28fffeb1-b58c-42d5-8219-c440d30248d9
#> 
#> Here is the R code:
#> 
#> ``` r
#> library("rjd3workspace")
#> 
#> jws <- jws_open("~/tmp.xml")
#> jsap <- jws_sap(jws, 1L)
#> sap_
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#653 ──
#> 
#> Title: Excel SA-Item names contain `\n` to separate sheet series
#> Text:
#> When we create a new SA-Item from an Excel file, the name of the series is automatically `sheet name\nseries name`.
#> 
#> On the GUI, it appears with an arrow ▶:
#> 
#> ![Image](https://github.com/user-attachments/assets/955bd1e0-98aa-4101-bc25-8af37974715b)
#> 
#> But on the XML file:
#> 
#> ![Image](https://github.com/user-attachments/asse
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#532 ──
#> 
#> Title: Unnecessary row displayed in Grid of time series
#> Text:
#> The following series has been imported from an Excel spreadsheet via the Providers panel. The series spans April 1982 to December 1985. However, when viewing the series in a Grid by double clicking on it, a row for 1986 is unnecessarily shown:
#> 
#> ![Image](https://github.com/user-attachments/assets/368a03b5-4e6f-44e1-bfeb
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#523 ──
#> 
#> Title: Add Copy Path/Reference context menu item to the Providers tab
#> Text:
#> The SDMX plugin uses a Copy Path/Reference context menu item to allow the user to quickly copy and paste relevant data IDs such as source, flow, and key. It is also used to create a command line with the appropriate parameters, allowing for some automation.
#> 
#> ![Image](https://github.com/user-attachments/assets/63c3066f-
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#345 ──
#> 
#> Title: Format type of date in output
#> Text:
#> Dear all, 
#> 
#> I noticed some trouble with the new output format: 
#> in version 2, the date was of format "General" in the output, however 
#> in version 3, the date is changed to a real date of format type: 
#> 1995-01-01T00:00:00/PM3, which creates problems in our subsequent macros. 
#> Is there any special reason behind thi
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#331 ──
#> 
#> Title: Improve SA output
#> Text:
#> Two problems with current implementation:
#> - the notification doesn't show the created file on Excel output
#> ![Screenshot 2024-06-14 133226](https://github.com/jdemetra/jdplus-main/assets/8778378/4664035b-69e5-4d18-a656-4f49a7104baa)
#> ![Screenshot 2024-06-14 133329](https://github.com/jdemetra/jdplus-main/assets/877837
#> ...
#> 
#> 
#> ── Issue jdemetra/jdplus-main#112 ──
#> 
#> Title: JWSACruncher and parameters `fullseriesname` and `refreshall`
#> Text:
#> I tried the following parameters files an a workspace with data created by an Excel file and it has no effect. Are the parameters `fullseriesname` and `refreshall` suppose to work in the JWSACruncher? Am I doing something wrong?
#> ```
#> <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
#> <wsaConfig bundle="10000" cs
#> ...
#>